==> code_supermarkets_france/data_acquisition/qlmc_2007_2012/build_qlmc_data/0_gen_ls_store_insee_matches.py <==
#!/usr/bin/python
# -*- coding: utf-8 -*-

import add_to_path
from add_to_path import path_data
import os, sys
import json

def dec_json(chemin):
  with open(chemin, 'r') as fichier:
    return json.loads(fichier.read())

def enc_json(database, chemin):
  with open(chemin, 'w') as fichier:
    json.dump(database, fichier)

ls_match = [[[u'SYSTEME U', u'BARALLE MARQUION'], ['BARALLE', '62860', 'PAS DE CALAIS', '62081']],
            [[u'E.LECLERC', u'MARLY'], ['MARLY', '57155', 'MOSELLE', '57447']],
            [[u'CORA', u'MASSY'], ['MASSY', '91300', 'ESSONNE', '91377']],
            [[u'AUCHAN', u'LE PONTET'], ['LE PONTET', '84130', 'VAUCLUSE', '84092']],
            [[u'GEANT', u'FONTAINE'], ['FONTAINE', '38600', 'ISERE', '38169']],
            [[u'GEANT', u'ST MARTIN DES CHAMPS'], ['ST MARTIN DES CHAMPS', '29600', 'FINISTERE', '29254']],
            [[u'CORA', u'STE MARGUERITE'], ['STE MARGUERITE', '88100', 'VOSGES', '88424']],
            [[u'CARREFOUR', u'PUGET SUR ARGENS CEDEX'], ['PUGET SUR ARGENS', '83480', 'VAR', '83099']],
            [[u'E.LECLERC', u'OLIVET'], ['OLIVET', '45160', 'LOIRET', '45232']],
            [[u'SYSTEME U', u'VERNOUILLET'], ['VERNOUILLET', '28500', 'EURE ET LOIR', '28404']],
            [[u'SYSTEME U', u'MONTBRISON'], ['MONTBRISON', '42600', 'LOIRE', '42147']],
            [[u'SYSTEME U', u'LES ARCS'], ['LES ARCS', '83460', 'VAR', '83004']],
            [[u'CARREFOUR', u'MONT SAINT AIGNAN'],['MONT ST AIGNAN', '76130', 'SEINE MARITIME', '76451']],
            [[u'CORA', u'PACE'], ['PACE', '35740', 'ILLE ET VILAINE', '35210']],
            [[u'GEANT', u'CHAMPNIERS'], ['CHAMPNIERS', '16430', 'CHARENTE', '16078']],
            [[u'AUCHAN', u'BIAS'], ['BIAS', '47300', 'LOT ET GARONNE', '47027']],
            [[u'CHAMPION', u'VERNON'], ['VERNON', '27200', 'EURE', '27681']],
            [[u'E.LECLERC', u'VALENCE'], ['VALENCE', '26000', 'DROME', '26362']],
            [[u'CORA', u'VERDUN'], ['VERDUN', '55100', 'MEUSE', '55545']],
            [[u'E.LECLERC', u'BOIS D ARCY'], ['BOIS D ARCY', '78390', 'YVELINES', '78073']],
            [[u'GEANT', u'AMILLY'], ['AMILLY', '45200', 'LOIRET', '45004']],
            [[u'INTERMARCHE', u'AMILLY'], ['AMILLY', '45200', 'LOIRET', '45004']],
            [[u'INTERMARCHE', u'ST CYPRIEN'], ['ST CYPRIEN', '42160', 'LOIRE', '42211']],
            [[u'E.LECLERC', u'VERDUN'], ['VERDUN', '55100', 'MEUSE', '55545']],
            [[u'SYSTEME U', u'LA MONTAGNE'], ['LA MONTAGNE', '44620', 'LOIRE ATLANTIQUE', '44101']],
            [[u'AUCHAN', u'BORDEAUX'], ['BORDEAUX', '33000', 'GIRONDE', '33063']],
            [[u'E.LECLERC', u'MARGON'], ['MARGON', '28400', 'EURE ET LOIR', '28236']],
            [[u'SYSTEME U', u'BASSENS'], ['BASSENS', '33530', 'GIRONDE', '33032']],
            [[u'E.LECLERC', u'ST GREGOIRE'], ['ST GREGOIRE', '35760', 'ILLE ET VILAINE', '35278']],
            [[u'AUCHAN', u'EPAGNY'], ['EPAGNY', '74330', 'HAUTE SAVOIE', '74112']],
            [[u'E.LECLERC', u'PRADINES'], ['PRADINES', '46090', 'LOT', '46224']],
            [[u'CHAMPION', u'ROCHEFORT'], ['ROCHEFORT', '17300', 'CHARENTE MARITIME', '17299']],
            [[u'SYSTEME U', u'CHATEAUBOURG'], ['CHATEAUBOURG', '35220', 'ILLE ET VILAINE', '35068']],
            [[u'CORA', u'ST MAXIMIN'], ['ST MAXIMIN', '60740', 'OISE', '60589']],
            [[u'CHAMPION', u'BRETEUIL'], ['BRETEUIL', '60120', 'OISE', '60104']], # second period
            [[u'INTERMARCHE', u'GIGNAC'], ['GIGNAC', '34150', 'HERAULT', '34114']],
            [[u'SUPER U', u'BOMPAS'], ['BOMPAS', '66430', 'PYRENEES ORIENTALES', '66021']],
            [[u'CENTRE E. LECLERC', u'ST ETIENNE FONTBELLON'], ['ST ETIENNE DE FONTBELLON', '07200', 'ARDECHE', '07231']],
            [[u'SUPER U', u'CRAON'], ['CRAON', '53400', 'MAYENNE', '53084']],
            [[u'SUPER U', u'DANNEMARIE'], ['DANNEMARIE', '68210', 'HAUT RHIN', '68068']],
            [[u'CENTRE E. LECLERC', u'BUXEROLLES'], ['BUXEROLLES', '86180', 'VIENNE', '86041']],
            [[u'AUCHAN', u'VILLARS'], ['VILLARS', '42390', 'LOIRE', '42330']],
            [[u'SUPER U', u'MARSEILLE RUE TADDEI'], ['MARSEILLE', '13000', 'BOUCHES DU RHONE', '13055']],
            [[u'CORA', u'LEMPDES'], ['LEMPDES', '63370', 'PUY DE DOME', '63193']],
            [[u'INTERMARCHE', u'LA CHARIT\xc9 SUR LOIRE'], ['LA CHARITE SUR LOIRE', '58400', 'NIEVRE', '58059']],
            [[u'CHAMPION', u'BOIS GUILLAUME AV EUROPE'], ['BOIS GUILLAUME', '76230', 'SEINE MARITIME', '76108']],
            [[u'CHAMPION', u'MOUTIERS TARENTAISE'], ['MOUTIERS', '73600', 'SAVOIE', '73181']],
            [[u'CENTRE E. LECLERC', u'ACHERES'], ['ACHERES', '78260', 'YVELINES', '78005']],
            [[u'CENTRE E. LECLERC', u'MONTDIDIER'], ['MONTDIDIER', '80500', 'SOMME', '80561']],
            [[u'GEANT CASINO', u'ST GREGOIRE'], ['ST GREGOIRE', '35760', 'ILLE ET VILAINE', '35278']],
            [[u'AUCHAN', u'TOURS NORD'], ['TOURS', '37000', 'INDRE ET LOIRE', '37261']],
            [[u'CARREFOUR', u'CHALON S. SAONE CC CHALON SUD'], ['CHALON SUR SAONE', '71100', 'SAONE ET LOIRE', '71076']],
            [[u'AUCHAN', u'CLAMECY'], ['CLAMECY', '58500', 'NIEVRE', '58079']],
            [[u'CENTRE E. LECLERC', u'PERPIGNAN LANGUEDOC'], ['PERPIGNAN', '66000', 'PYRENEES ORIENTALES', '66136']],
            [[u'AUCHAN', u'FAYET'], ['FAYET', '02100', 'AISNE', '02303']],
            [[u'CENTRE E. LECLERC', u'NORMANVILLE'], ['NORMANVILLE', '27930', 'EURE', '27439']],
            [[u'HYPER U', u'VERNOUILLET'], ['VERNOUILLET', '28500', 'EURE ET LOIR', '28404']], # duplicate SYSTEME U?
            [[u'AUCHAN', u'BLOIS VINEUIL'], ['VINEUIL', '41350', 'LOIR ET CHER', '41295']],
            [[u'CARREFOUR', u'BASSENS'], ['BASSENS', '73000', 'SAVOIE', '73031']],
            [[u'AUCHAN', u'MAUREPAS'], ['MAUREPAS', '78310', 'YVELINES', '78383']],
            [[u'CARREFOUR', u'NICE N202'], ['NICE', '06000', 'ALPES MARITIMES', '06088']],
            [[u'CENTRE E. LECLERC', u'LES ANGLES'], ['LES ANGLES', '30133', 'GARD', '30011']],
            [[u'GEANT CASINO', u'FENOUILLET'], ['FENOUILLET', '31150', 'HAUTE GARONNE', '31182']],
            [[u'GEANT CASINO', u'CASTRES'], ['CASTRES', '81100', 'TARN', '81065']],
            [[u'INTERMARCHE', u'GRANDVILLIERS'], ['GRANDVILLIERS', '60210', 'OISE', '60286']],
            [[u'GEANT CASINO', u'CHAUMONT'], ['CHAUMONT', '52000', 'HAUTE MARNE', '52121']],
            [[u'CENTRE E. LECLERC', u'LURE'], ['LURE', '70200', 'HAUTE SAONE', '70310']],
            [[u'CENTRE E. LECLERC', u'MASSY'], ['MASSY', '91300', 'ESSONNE', '91377']],
            [[u'INTERMARCHE', u'LE PASSAGE'], ['LE PASSAGE', '47520', 'LOT ET GARONNE', '47201']],
            [[u'CENTRE E. LECLERC', u'CHAMBRY'], ['CHAMBRY', '02000', 'AISNE', '02157']],
            [[u'INTERMARCHE', u'EVREUX AV 14 JUILLET'], ['EVREUX', '27000', 'EURE', '27229']],
            [[u'CENTRE E. LECLERC', u'ST PIERRE DOLERON'], ['ST PIERRE D OLERON', '17310', 'CHARENTE MARITIME', '17385']],
            [[u'INTERMARCHE', u'USSEL'], ['USSEL', '19200', 'CORREZE', '19275']],
            [[u'GEANT CASINO', u'FONTAINE'], ['FONTAINE', '38600', 'ISERE', '38169']], # duplicate GEANT or confusion w/ FONTAINE LES DIJON
            [[u'GEANT CASINO', u'ST LOUIS'], ['ST LOUIS', '68300', 'HAUT RHIN', '68297']],
            [[u'CHAMPION', u'FRENEUSE'], ['FRENEUSE', '78840', 'YVELINES', '78255']],
            [[u'GEANT CASINO', u'AVIGNON CC CAP SUD'], ['AVIGNON', '84000', 'VAUCLUSE', '84007']],
            [[u'SUPER U', u'CASTRES'], ['CASTRES', '81100', 'TARN', '81065']],
            [[u'AUCHAN', u'MONT ST MARTIN'], ['MONT ST MARTIN', '54350', 'MEURTHE ET MOSELLE', '54382']], # third period
            [[u'AUCHAN', u'CLERMONT EN AUVERGNE'], ['CLERMONT FERRAND', '63100', 'PUY DE DOME', '63113']],
            [[u'GEANT CASINO', u'CLERMONT FERRAND BREZET'], ['CLERMONT FERRAND', '63100', 'PUY DE DOME', '63113']],
            [[u'SUPER U', u'ROCHE SUR YON'], ['LA ROCHE SUR YON', '85000', 'VENDEE', '85191']],
            [[u'GEANT CASINO', u'VALENCE'], ['VALENCE', '26000', 'DROME', '26362']],
            [[u'AUCHAN', u'BORDEAUX C.C. MERIADECK'], ['BORDEAUX', '33000', 'GIRONDE', '33063']],
            [[u'INTERMARCHE', u'ROUEN RUE CARREL'], ['ROUEN', '76000', 'SEINE MARITIME', '76540']],
            [[u'SUPER U', u'BOURGUET'], []], # mistake Bourget? Maubourguet? => rather drop it
            [[u'GEANT CASINO', u'AMILLY'], ['AMILLY', '45200', 'LOIRET', '45004']], # duplicate GEANT?
            [[u'CENTRE E. LECLERC', u"RAON L'ETAPE"], ['RAON L ETAPE', '88110', 'VOSGES', '88372']],
            [[u'AUCHAN', u'LA TRINITE'], ['LA TRINITE', '06340', 'ALPES MARITIMES', '06149']],
            [[u'GEANT CASINO', u'TORCY'], ['TORCY', '71210', 'SAONE ET LOIRE', '71540']],
            [[u'CHAMPION', u'PARIS 13 NATIONALE'], ['PARIS', '75000', 'PARIS', '75056']],
            [[u'CENTRE LECLERC', u'CLERMONT FERR. FLAUBERT'], ['CLERMONT FERRAND', '63000', 'PUY DE DOME', '63113']],
            [[u'GEANT CASINO', u'CARCASSONNE C.C. CITE2'], ['CARCASSONNE', '11000', 'AUDE', '11069']],
            [[u'SUPER U', u'LE PONT DE BEAUVOISIN'], ['LE PONT DE BEAUVOISIN', '73330', 'SAVOIE', '73204']],
            [[u'INTERMARCHE', u'VITRE'], ['VITRE', '35500', 'ILLE ET VILAINE', '35360']],
            [[u'CHAMPION', u'CLERMONT FERRAND ST ALYRE'], ['CLERMONT FERRAND', '63000', 'PUY DE DOME', '63113']],
            [[u'SUPER U', u'MAGNE'], ['MAGNE', '79460', 'DEUX SEVRES', '79162']],
            [[u'INTERMARCHE', u'ROUEN RUE PARC'], ['ROUEN', '76000', 'SEINE MARITIME', '76540']],
            [[u'CHAMPION', u'LIMOGES MAS LOUBIER'], ['LIMOGES', '87000', 'HAUTE VIENNE', '87085']],
            [[u'SUPER U', u'CAEN BEAULIEU'], ['CAEN', '14000', 'CALVADOS', '14118']],
            [[u'CHAMPION', u'AUBUSSON'], ['AUBUSSON', '23200', 'CREUSE', '23008']],
            [[u'CHAMPION', u'REIMS AV EPERNAY'], ['REIMS', '51100', 'MARNE', '51454']],
            [[u'INTERMARCHE', u'POLIGNY'], ['POLIGNY', '39800', 'JURA', '39434']],
            [[u'CHAMPION', u'CLERMONT FERRAND BARR. DE JAUD'],  ['CLERMONT FERRAND', '63000', 'PUY DE DOME', '63113']],
            [[u'CHAMPION', u'BOIS GUILLAUME REPUBLIQUE'], ['BOIS GUILLAUME', '76230', 'SEINE MARITIME', '76108']],
            [[u'CHAMPION', u'NIORT ANGELY'], ['NIORT', '79000', 'DEUX SEVRES', '79191']],
            [[u'CHAMPION', u'LIMOGES PAGUENAUD'], ['LIMOGES', '87000', 'HAUTE VIENNE', '87085']],
            [[u'CHAMPION', u'ARQUES'], ['ARQUES', '62510', 'PAS DE CALAIS', '62040']],
            [[u'CARREFOUR', u'STE GENEVIEVE DES BOIS'], ['STE GENEVIEVE DES BOIS', '91700', 'ESSONNE', '91549']],
            [[u'AUCHAN', u'DURY'], ['DURY', '80480', 'SOMME', '80261']],
            [[u'CHAMPION', u'COZE'], ['COZES', '17120', 'CHARENTE MARITIME', '17131']], #mispelled
            [[u'CENTRE E. LECLERC', u'NICE RTE TURIN'], ['NICE', '06300', 'ALPES MARITIMES', '06088']],
            [[u'CHAMPION', u'ST FLOUR'], ['ST FLOUR', '15100', 'CANTAL', '15187']],
            [[u'INTERMARCHE', u"NIORT ROUTE D'AIFFRES"], ['NIORT', '79000', 'DEUX SEVRES', '79191']],
            [[u'INTERMARCHE', u'CHELLES'], ['CHELLES', '77500', 'SEINE ET MARNE', '77108']],
            [[u'CENTRE E. LECLERC', u'LILLE FIVES'], ['LILLE', '59800', 'NORD', '59350']],
            [[u'CENTRE E. LECLERC', u'ST LOUP'], ['ST LOUP', '69490', 'RHONE', '69223']],
            [[u'INTERMARCHE', u'MONTEILS'], ['MONTEILS', '82300', 'TARN ET GARONNE', '82126']],
            [[u'GEANT CASINO', u'NIORT CHAURAY'], ['CHAURAY', '79180', 'DEUX SEVRES', '79081']],
            [[u'SUPER U', u'FRANQUEVILLE SAINT PIERRE'], ['FRANQUEVILLE ST PIERRE', '76520', 'SEINE MARITIME', '76475']],
            [[u'AUCHAN', u'CLAMECY'], ['CLAMECY', '58500', 'NIEVRE', '58079']],
            [[u'CENTRE E. LECLERC', u'FLERS'], ['FLERS', '61100', 'ORNE', '61169']],
            [[u'CHAMPION', u'BOIS GUILLAUME BOCQUETS'], ['BOIS GUILLAUME', '76230', 'SEINE MARITIME', '76108']], # duplicate av europe
            [[u'SUPER U', u'GRENADE SUR GARONNE'], ['GRENADE', '31330', 'HAUTE GARONNE', '31232']],
            [[u'CHAMPION', u'NIORT AV. PARIS'], ['NIORT', '79000', 'DEUX SEVRES', '79191']],
            [[u'INTERMARCHE', u'CLERMONT FERRAND FONTGIEVE'], ['CLERMONT FERRAND', '63000', 'PUY DE DOME', '63113']],
            [[u'INTERMARCHE', u'ST AMBROIX'], ['ST AMBROIX', '30500', 'GARD', '30227']],
            [[u'CENTRE E. LECLERC', u'BETTING'], ['FREYMING MERLEBACH', '57800', 'MOSELLE', '57240']],
            [[u'AUCHAN', u'CASTRES'], ['CASTRES', '81100', 'TARN', '81065']],
            [[u'CHAMPION', u'TOURS 501E REGT INFANTERIE'], ['TOURS', '37000', 'INDRE ET LOIRE', '37261']], # fourth period
            [[u'INTERMARCHE', u'ST NAZAIRE'], ['ST NAZAIRE', '44600', 'LOIRE ATLANTIQUE', '44184']], # several st nazaire but...
            [[u'CHAMPION', u'La Crau'], ['LA CRAU', '83260', 'VAR', '83047']],
            [[u'CHAMPION', u'LILLE FIVES'], ['LILLE', '59800', 'NORD', '59350']],
            [[u'INTERMARCHE', u'MOULINS'], ['MOULINS', '03000', 'ALLIER', '03190']],
            [[u'SUPER U', u"L'ISLE JOURDAIN"], ['L ISLE JOURDAIN', '32600', 'GERS', '32160']],
            [[u'CENTRE E. LECLERC', u'BERGERAC PASTEUR'], ['BERGERAC', '24100', 'DORDOGNE', '24037']],
            [[u'CHAMPION', u'MARSEILLE 12 AMARYLLIS'], ['MARSEILLE 12EME ARRONDISS', '13012', 'BOUCHES DU RHONE', '13212']],
            [[u'SUPER U', u'SAINT MAXIMIN'], ['ST MAXIMIN LA STE BAUME', '83470', 'VAR', '83116']], # not clear vs. st maximin
            [[u'INTERMARCHE', u'ETRECHY'], ['ETRECHY', '91580', 'ESSONNE', '91226']], # several
            [[u'INTERMARCHE', u'ST LOUIS'], ['ST LOUIS', '68300', 'HAUT RHIN', '68297']],
            [[u'INTERMARCHE', u'ROCHEFORT'], ['ROCHEFORT', '17300', 'CHARENTE MARITIME', '17299']],
            [[u'SUPER U', u'ROUILLAC'], ['ROUILLAC', '16170', 'CHARENTE', '16286']],
            [[u'CENTRE E. LECLERC', u'LAVAL'], ['LAVAL', '53000', 'MAYENNE', '53130']],
            [[u'SUPER U', u'MONTREUIL'], ['MONTREUIL', '93100', 'SEINE SAINT DENIS', '93048']],
            [[u'CENTRE E. LECLERC', u'ERSTEIN CEDEX'], ['ERSTEIN', '67150', 'BAS RHIN', '67130']],
            [[u'CHAMPION', u"LES SABLES D'OLONNE GARLIERE"],  ['LES SABLES D OLONNE', '85100', 'VENDEE', '85194']],
            [[u'CHAMPION', u'ST CYPRIEN'], ['ST CYPRIEN', '24220', 'DORDOGNE', '24396']],
            [[u'HYPER U', u'RUMILLY'], ['RUMILLY', '74150', 'HAUTE SAVOIE', '74225']],
            [[u'SUPER U', u'PRADES'], ['PRADES', '66500', 'PYRENEES ORIENTALES', '66149']],
            [[u'SUPER U', u'STE GENEVIEVE'], ['STE GENEVIEVE', '60730', 'OISE', '60575']],
            [[u'CHAMPION', u'TOURS LES FONTAINES'], ['TOURS', '37200', 'INDRE ET LOIRE', '37261']],
            [[u'SUPER U', u'EGUZON CHANTOME'], ['EGUZON-CHANTOME', '36270', 'INDRE', '36070']],
            [[u'SUPER U', u'ILLARS LES DOMBES'], ['VILLARS LES DOMBES', '01330', 'AIN', '01443']], # mispelled
            [[u'CHAMPION', u'ECUELLES'], ['ECUELLES', '77250', 'SEINE ET MARNE', '77166']],
            [[u'SUPER U', u'CAMBLANES'], ['CAMBLANES ET MEYNAC', '33360', 'GIRONDE', '33085']],
            [[u'INTERMARCHE', u'BORDES'], ['BORDES', '64510', 'PYRENEES ATLANTIQUES', '64138']],
            [[u'INTERMARCHE', u'CARCASSONNE PORTES CAMARGUE'], ['CARCASSONNE', '11000', 'AUDE', '11069']],
            [[u'CHAMPION', u'ST GRATIEN'], ['ST GRATIEN', '95210', "VAL D'OISE", '95555']],
            [[u'CHAMPION', u'RAILLENCOURT SAINTE OLLE'], ['RAILLENCOURT STE OLLE', '59554', 'NORD', '59488']],
            [[u'SUPER U', u'MARSEILLE 5 SAKAKINI'], ['MARSEILLE 5EME ARRONDISSE', '13005', 'BOUCHES DU RHONE', '13205']],
            [[u'CHAMPION', u'NEVERS COLBERT'], ['NEVERS', '58000', 'NIEVRE', '58194']],
            [[u'SUPER U', u'PONT DE BEAUVOISIN'], ['LE PONT DE BEAUVOISIN', '73330', 'SAVOIE', '73204']],
            [[u'CHAMPION', u'COURBEVOIE CHARRAS'], ['COURBEVOIE', '92400', 'HAUTS DE SEINE', '92026']],
            [[u'SUPER U', u'NOZAY'], ['NOZAY', '44170', 'LOIRE ATLANTIQUE', '44113']],
            [[u'GEANT CASINO', u'ST NAZAIRE'], ['ST NAZAIRE', '44600', 'LOIRE ATLANTIQUE', '44184']], #fifth period
            [[u'HYPER CHAMPION', u'SAUSSAY'], ['SAUSSAY', '28260', 'EURE ET LOIR', '28371']],
            [[u'SUPER U', u'ANGERS L. JOUVET'], ['ANGERS', '49000', 'MAINE ET LOIRE', '49007']],
            [[u'LECLERC', u'ST RAPHAEL'], ['ST RAPHAEL', '83700', 'VAR', '83118']],
            [[u'INTERMARCHE', u'EVREUX FAUVILLE'], ['EVREUX', '27000', 'EURE', '27229']],
            [[u'INTERMARCHE', u'CLERMONT'],  ['CLERMONT', '60600', 'OISE', '60157']], # not absolutely sure
            [[u'GEANT CASINO', u'CARCASSONNE SOUV. Fran\xe7ais'], ['CARCASSONNE', '11000', 'AUDE', '11069']],
            [[u'CARREFOUR', u'GOUSSAINVILLE'], ['GOUSSAINVILLE', '95190', "VAL D'OISE", '95280']],
            [[u'LECLERC', u'CLAMECY'], ['CLAMECY', '58500', 'NIEVRE', '58079']],
            [[u'INTERMARCHE', u'MURAT'], ['MURAT', '15300', 'CANTAL', '15138']],
            [[u'LECLERC', u'VITRE'], ['VITRE', '35500', 'ILLE ET VILAINE', '35360']],
            [[u'INTERMARCHE', u'BERNAY GOMBERT'], ['BERNAY', '27300', 'EURE', '27056']],
            [[u'AUCHAN', u'FLOIRAC'], ['FLOIRAC', '33270', 'GIRONDE', '33167']], # or ('BOULIAC', '33270', 'GIRONDE', '33065')
            [[u'LECLERC', u'SAINT DIZIER'], ['ST DIZIER', '52100', 'HAUTE MARNE', '52448']],
            [[u'SUPER U', u'L ISLE JOURDAIN'], ['L ISLE JOURDAIN', '32600', 'GERS', '32160']],
            [[u'SUPER U', u'NEUVILLE AUX BOIS'], ['NEUVILLE AUX BOIS', '45170', 'LOIRET', '45224']],
            [[u'CHAMPION', u'CIVRAY'], ['CIVRAY', '86400', 'VIENNE', '86078']],
            [[u'CARREFOUR MARKET', u'MARLY'], ['MARLY', '59770', 'NORD', '59383']], # not 100% sure (marly moselle or..)
            [[u'LECLERC', u'SAINT LUBIN DES JONCHERETS'], ['ST LUBIN DES JONCHERETS', '28350', 'EURE ET LOIR', '28348']],
            [[u'GEANT CASINO', u'ANGERS CC ANJOU'], ['ANGERS', '49000', 'MAINE ET LOIRE', '49007']],
            [[u'LECLERC', u"RAON L'ETAPE"], ['RAON L ETAPE', '88110', 'VOSGES', '88372']],
            [[u'CHAMPION', u'PARIS 13 CC Italie 2'], ['PARIS 13EME ARRONDISSEMENT', '75013', 'PARIS', '75113']],
            [[u'GEANT CASINO', u'ANGERS LA ROSERAIE'], ['ANGERS', '49000', 'MAINE ET LOIRE', '49007']],
            [[u'LECLERC', u'LA SEYNE SUR MER LES SABLETTES'], ['LA SEYNE SUR MER', '83500', 'VAR', '83126']],
            [[u'CHAMPION', u'ST PRIEST'], ['ST PRIEST', '69800', 'RHONE', '69290']],
            [[u'AUCHAN', u'OLIVET'], ['OLIVET', '45160', 'LOIRET', '45232']],
            [[u'CHAMPION', u'PESSAC MORIN'], ['PESSAC', '33600', 'GIRONDE', '33318']],
            [[u'INTERMARCHE', u'ST FLOUR'], ['ST FLOUR', '15100', 'CANTAL', '15187']],
            [[u'CHAMPION', u'PESSAC DULOUT'], ['PESSAC', '33600', 'GIRONDE', '33318']],
            [[u'CHAMPION', u'NIORT PARIS'], ['NIORT', '79000', 'DEUX SEVRES', '79191']], 
            [[u'SUPER U', u'ANGERS BOURG DE PAILLE'], ['BEAUCOUZE', '49070', 'MAINE ET LOIRE', '49020']], # suburb of Angers
            [[u'LECLERC', u'ST LOUIS'], ['ST LOUIS', '68300', 'HAUT RHIN', '68297']],
            [[u'INTERMARCHE', u'BERNAY THIERS'], ['BERNAY', '27300', 'EURE', '27056']],
            [[u'CARREFOUR MARKET', u'RENNES POTERIE'], ['RENNES', '35200', 'ILLE ET VILAINE', '35238']],
            [[u'AUCHAN', u'ST PRIEST'], ['ST PRIEST', '69800', 'RHONE', '69290']],
            [[u'SUPER U', u'PARIS 17'], ['PARIS 17EME ARRONDISSEMENT', '75017', 'PARIS', '75117']],
            [[u'SUPER U', u'CHATEAUBOURG'], ['CHATEAUBOURG', '35220', 'ILLE ET VILAINE', '35068']],
            [[u'GEANT CASINO', u'MONTPELLIER ARGELLIERS'], ['MONTPELLIER', '34000', 'HERAULT', '34172']],
            [[u'HYPER CHAMPION', u'ST NAZAIRE'], ['ST NAZAIRE', '44600', 'LOIRE ATLANTIQUE', '44184']],
            [[u'CARREFOUR MARKET', u'VILLEVIEILLE'], ['VILLEVIEILLE', '30250', 'GARD', '30352']],
            [[u'CARREFOUR MARKET', u'RENNES CC BOURG EVEQUE'],  ['RENNES', '35000', 'ILLE ET VILAINE', '35238']],
            [[u'CHAMPION', u'REIMS NEUFCHATEL'], ['REIMS', '51100', 'MARNE', '51454']],
            [[u'HYPER CHAMPION', u'ST SULPICE'], ['ST SULPICE', '81370', 'TARN', '81271']],
            [[u'SUPER U', u'BROU'], ['BROU', '28160', 'EURE ET LOIR', '28061']],
            [[u'LECLERC', u'ST GREGOIRE'], ['ST GREGOIRE', '35760', 'ILLE ET VILAINE', '35278']],
            [[u'GEANT CASINO', u'MONTPELLIER LODEVE'], ['MONTPELLIER', '34000', 'HERAULT', '34172']], # 34070?
            [[u'CARREFOUR', u'MERIGNAC'], ['MERIGNAC', '33700', 'GIRONDE', '33281']],
            [[u'CHAMPION', u'ST MARCEL'], ['ST MARCEL', '71380', 'SAONE ET LOIRE', '71445']],
            [[u'CARREFOUR', u'LIEUSAINT'], ['LIEUSAINT', '77127', 'SEINE ET MARNE', '77251']],
            [[u'CHAMPION', u"LA CHAUSSEE D'IVRY"], ['LA CHAUSSEE D IVRY', '28260', 'EURE ET LOIR', '28096']],
            [[u'GEANT CASINO', u'PARIS CC MASSENA 13'], ['PARIS 13EME ARRONDISSEMENT', '75013', 'PARIS', '75113']],
            [[u'SUPER U', u'VILLE'], ['VILLE', '67220', 'BAS RHIN', '67507']],
            [[u'INTERMARCHE', u'ST FLORENTIN'], ['ST FLORENTIN', '89600', 'YONNE', '89345']],
            [[u'CHAMPION', u'PARIS 18 BARBES'], ['PARIS 18EME ARRONDISSEMENT', '75018', 'PARIS', '75118']],
            [[u'CARREFOUR', u'ST DENIS'], ['ST DENIS', '93200', 'SEINE SAINT DENIS', '93066']],
            [[u'SUPER U', u'VILLIERS SUR MARNE'], ['VILLIERS SUR MARNE', '94350', 'VAL DE MARNE', '94079']],
            [[u'AUCHAN', u'BORDEAUX CC LE LAC'], ['BORDEAUX', '33300', 'GIRONDE', '33063']], # 33000?
            [[u'CHAMPION', u'REIMS CLEMENCEAU'], ['REIMS', '51100', 'MARNE', '51454']],
            [[u'CHAMPION', u'MAURIAC'], ['MAURIAC', '15200', 'CANTAL', '15120']],
            [[u'GEANT CASINO', u'CARCASSONNE CC SALVAZA'], ['CARCASSONNE', '11000', 'AUDE', '11069']],
            [[u'GEANT CASINO', u'VALENCE CHANTECOURIOL'], ['VALENCE', '26000', 'DROME', '26362']],
            [[u'AUCHAN', u'BORDEAUX CC MERIADECK'], ['BORDEAUX', '33000', 'GIRONDE', '33063']],
            [[u'SUPER U', u'VALMONT'], ['VALMONT', '57730', 'MOSELLE', '57690']],
            [[u'CHAMPION', u'PARIS 14 GAL LECLERC'], ['PARIS 14EME ARRONDISSEMENT', '75014', 'PARIS', '75114']],
            [[u'INTERMARCHE', u'ST MARCEL'], ['ST MARCEL', '27950', 'EURE', '27562']],
            [[u'LECLERC', u'LA SEYNE SUR MER LERY'], ['LA SEYNE SUR MER', '83500', 'VAR', '83126']],
            [[u'CHAMPION', u'DESERTINES'], ['DESERTINES', '03630', 'ALLIER', '03098']],
            [[u'LECLERC', u'LE BREUIL'], ['LE BREUIL', '71670', 'SAONE ET LOIRE', '71059']], 
            [[u'LECLERC', u'VERDUN'], ['VERDUN', '55100', 'MEUSE', '55545']],
            [[u'CARREFOUR MARKET', u'NIORT ANGELY'], ['NIORT', '79000', 'DEUX SEVRES', '79191']],
            [[u'AUCHAN', u'VINEUIL'], ['VINEUIL', '41350', 'LOIR ET CHER', '41295']],
            [[u'INTERMARCHE', u'BOOS'], ['BOOS', '76520', 'SEINE MARITIME', '76116']],
            [[u'CHAMPION', u'BERNAY'], ['BERNAY', '27300', 'EURE', '27056']],
            [[u'GEANT CASINO', u'VALENCE CC VALENCE 2'], ['VALENCE', '26000', 'DROME', '26362']],
            [[u'INTERMARCHE', u'PRADES'], ['PRADES', '66500', 'PYRENEES ORIENTALES', '66149']],
            [[u'INTERMARCHE', u'VERNOUILLET'], ['VERNOUILLET', '28500', 'EURE ET LOIR', '28404']],
            [[u'SUPER U', u'CHARLY SUR MARNE'], ['CHARLY', '02310', 'AISNE', '02163']],
            [[u'LECLERC', u'ST DENIS'], ['ST DENIS', '89100', 'YONNE', '89342']], # not 100% sure?
            [[u'INTERMARCHE', u'REVEL'], ['REVEL', '31250', 'HAUTE GARONNE', '31451']],
            [[u'CARREFOUR MARKET', u'SEYSSEL'], ['SEYSSEL', '74910', 'HAUTE SAVOIE', '74269']],
            [[u'GEANT CASINO', u'MARSEILLE 11'], ['MARSEILLE 11EME ARRONDISS', '13011', 'BOUCHES DU RHONE', '13211']],
            [[u'GEANT CASINO', u'MARSEILLE 12'], ['MARSEILLE 12EME ARRONDISS', '13012', 'BOUCHES DU RHONE', '13212']],
            [[u'GEANT CASINO', u'ST MARTIN DES CHAMPS'], ['ST MARTIN DES CHAMPS', '29600', 'FINISTERE', '29254']],
            [[u'SUPER U', u'BUXEROLLES'], ['BUXEROLLES', '86180', 'VIENNE', '86041']],
            [[u'GEANT CASINO', u'MARSEILLE 9'], ['MARSEILLE 9EME ARRONDISSE', '13009', 'BOUCHES DU RHONE', '13209']],
            [[u'CHAMPION', u'ST MAURICE'], ['ST MAURICE', '94410', 'VAL DE MARNE', '94069']],
            [[u'LECLERC', u'ST MAUR'], ['ST MAUR', '36250', 'INDRE', '36202']],
            [[u'INTERMARCHE', u'EVREUX 14 JUILLET'], ['EVREUX', '27000', 'EURE', '27229']],
            [[u'INTERMARCHE', u'ST JEAN LE BLANC'], ['ST JEAN LE BLANC', '45650', 'LOIRET', '45286']], # sixth period
            [[u'INTERMARCHE', u'BELLEVILLE'], ['BELLEVILLE', '69220', 'RHONE', '69019']], # not 100%
            [[u'CARREFOUR MARKET', u'CHATEAUROUX'], ['CHATEAUROUX', '36000', 'INDRE', '36044']],
            [[u'CHAMPION', u'AUTERIVE'], ['AUTERIVE', '31190', 'HAUTE GARONNE', '31033']],
            [[u'INTERMARCHE', u'RUMILLY'], ['RUMILLY', '74150', 'HAUTE SAVOIE', '74225']],
            [[u'CARREFOUR MARKET', u'TOULOUSE RUE ROUGENET'], ['TOULOUSE', '31500', 'HAUTE GARONNE', '31555']],
            [[u'LECLERC', u'VERNON'], ['VERNON', '27200', 'EURE', '27681']],
            [[u'LECLERC', u'CHARMES'], ['CHARMES', '88130', 'VOSGES', '88090']],
            [[u'SUPER U', u'MARSEILLE 5 ST PIERRE'], ['MARSEILLE 5EME ARRONDISSE', '13005', 'BOUCHES DU RHONE', '13205']],
            [[u'CARREFOUR MARKET', u'TOULOUSE RTE SEYSSES'], ['TOULOUSE', '31000', 'HAUTE GARONNE', '31555']],
            [[u'LECLERC', u"L'ISLE ADAM"], ['L ISLE ADAM', '95290', "VAL D'OISE", '95313']],
            [[u'CHAMPION', u'REIMS RUE GAMBETTA'], ['REIMS', '51100', 'MARNE', '51454']],
            [[u'CHAMPION', u'POITIERS PLACE PROVENCE'], ['POITIERS', '86000', 'VIENNE', '86194']],
            [[u'INTERMARCHE', u'SENLIS'],  ['SENLIS', '60300', 'OISE', '60612']],
            [[u'CARREFOUR MARKET', u'THONON L. BAINS LANDAZ'], ['THONON LES BAINS', '74200', 'HAUTE SAVOIE', '74281']],
            [[u'GEANT CASINO', u'PARIS 13 MASSENA'], ['PARIS 13EME ARRONDISSEMENT', '75013', 'PARIS', '75113']],
            [[u'LECLERC', u"ST OUEN L'AUMONE"], ['ST OUEN L AUMONE', '95310', "VAL D'OISE", '95572']],
            [[u'SUPER U', u'PERREUX'], ['PERREUX', '42120', 'LOIRE', '42170']],
            [[u'CHAMPION', u'NEVERS AV COLBERT'], ['NEVERS', '58000', 'NIEVRE', '58194']],
            [[u'GEANT CASINO', u"VILLENAVE D'ORNON"], ['VILLENAVE D ORNON', '33140', 'GIRONDE', '33550']],
            [[u'LECLERC', u'COLOMBES RUE MICHELET'], ['COLOMBES', '92700', 'HAUTS DE SEINE', '92025']],
            [[u'LECLERC', u'BEAUNE'], ['BEAUNE', '21200', "COTE D'OR", '21054']],
            [[u'SUPER U', u'EYSSAC'], ['MEYSSAC', '19500', 'CORREZE', '19138']],
            [[u'CARREFOUR MARKET', u"L'ISLE JOURDAIN"], ['L ISLE JOURDAIN', '32600', 'GERS', '32160']],
            [[u'LECLERC', u'RIBEAUVILLE'], ['RIBEAUVILLE', '68150', 'HAUT RHIN', '68269']],
            [[u'LECLERC', u'OLIVET'], ['OLIVET', '45160', 'LOIRET', '45232']],
            [[u'INTERMARCHE', u'AURILLAC AV GAL LECLERC'], ['AURILLAC', '15000', 'CANTAL', '15014']],
            [[u'GEANT CASINO', u'MARSEILLE 8 MICHELET'], ['MARSEILLE 8EME ARRONDISSE', '13008', 'BOUCHES DU RHONE', '13208']],
            [[u'CARREFOUR MARKET', u'BERNAY'], ['BERNAY', '27300', 'EURE', '27056']],
            [[u'GEANT CASINO', u'MONTPELLIER AV ARGELLIERS'], ['MONTPELLIER', '34000', 'HERAULT', '34172']],
            [[u'SUPER U', u'LAVAUR'], ['LAVAUR', '81500', 'TARN', '81140']],
            [[u'LECLERC', u'STE EULALIE'], ['STE EULALIE', '33560', 'GIRONDE', '33397']],
            [[u'SUPER U', u"L'ARBRESLE"], ['L ARBRESLE', '69210', 'RHONE', '69010']],
            [[u'CARREFOUR', u'MONDEVILLE'], ['MONDEVILLE', '14120', 'CALVADOS', '14437']],
            [[u'SUPER U', u"LA MURE D'ISERE"], ['LA MURE', '38350', 'ISERE', '38269']],
            [[u'SUPER U', u"ST SYLVAIN D'ANJOU"], ['ST SYLVAIN D ANJOU', '49480', 'MAINE ET LOIRE', '49323']],
            [[u'CARREFOUR MARKET', u'THONON L. BAINS FERRY'], ['THONON LES BAINS', '74200', 'HAUTE SAVOIE', '74281']],
            [[u'GEANT CASINO', u'MARSEILLE 11 LA VALENTINE'], ['MARSEILLE 11EME ARRONDISS', '13011', 'BOUCHES DU RHONE', '13211']],
            [[u'AUCHAN', u"VILLENEUVE D'ASCQ"], ['VILLENEUVE D ASCQ', '59650', 'NORD', '59009']],
            [[u'SUPER U', u'ST GEORGES BAILLARGEAUX'], ['ST GEORGES LES BAILLARGEA', '86130', 'VIENNE', '86222']],
            [[u'LECLERC', u'BERGERAC RTE BORDEAUX'], ['BERGERAC', '24100', 'DORDOGNE', '24037']],
            [[u'CHAMPION', u"ST MARTIN D'AUXIGNY"], ['ST MARTIN D AUXIGNY', '18110', 'CHER', '18223']],
            [[u'LECLERC', u'EXPRESS MONTFORT SUR MEU'], ['MONTFORT SUR MEU', '35160', 'ILLE ET VILAINE', '35188']], # DROP ONCE UPDATE DONE
            [[u'LECLERC EXPRESS', u'MONTFORT SUR MEU'], ['MONTFORT SUR MEU', '35160', 'ILLE ET VILAINE', '35188']], # KEEP (if necessary?)
            [[u'LECLERC', u'COLOMBES BLD DE GAULLE'], ['COLOMBES', '92700', 'HAUTS DE SEINE', '92025']],
            [[u'INTERMARCHE', u'CROIX'], ['CROIX', '59170', 'NORD', '59163']],
            [[u'HYPER U', u"CLERMONT L'HERAULT"], ['CLERMONT L HERAULT', '34800', 'HERAULT', '34079']],
            [[u'SUPER U', u"L'EPINE"], ['L EPINE', '85740', 'VENDEE', '85083']], # probably disappeared
            [[u'INTERMARCHE', u'RENNES CC LONGS CHAMPS'], ['RENNES', '35700', 'ILLE ET VILAINE', '35238']],
            [[u'CARREFOUR MARKET', u'PIERRES'], ['PIERRES', '28130', 'EURE ET LOIR', '28298']],
            [[u'INTERMARCHE', u'ROYE'], ['ROYE', '80700', 'SOMME', '80685']],
            [[u'CORA', u"VILLENEUVE D'ASCQ"], ['VILLENEUVE D ASCQ', '59650', 'NORD', '59009']],
            [[u'CARREFOUR', u'MONTREUIL'], ['MONTREUIL', '93100', 'SEINE SAINT DENIS', '93048']],
            [[u'CARREFOUR MARKET', u'GRISOLLES'], ['GRISOLLES', '82170', 'TARN ET GARONNE', '82075']],
            [[u'GEANT CASINO', u'MONTPELLIER AV LODEVE'], ['MONTPELLIER', '34070', 'HERAULT', '34172']],
            [[u'INTERMARCHE', u'AURILLAC RUE FIRMINY'], ['AURILLAC', '15000', 'CANTAL', '15014']],
            [[u'CHAMPION', u'POITIERS RUE MERMOZ'], ['POITIERS', '86000', 'VIENNE', '86194']],
            [[u'INTERMARCHE', u'MAINVILLIERS'], ['MAINVILLIERS', '28300', 'EURE ET LOIR', '28229']],
            [[u'CHAMPION', u'PARIS 17 CLIGNANCOURT'], ['PARIS 17EME ARRONDISSEMENT', '75017', 'PARIS', '75117']],
            [[u'AUCHAN', u"MARSAC SUR L'ISLE"], ['MARSAC SUR L ISLE', '24430', 'DORDOGNE', '24256']],
            [[u'SUPER U', u'MARSEILLE 7 TADDEI'], ['MARSEILLE 7EME ARRONDISSE', '13007', 'BOUCHES DU RHONE', '13207']],
            [[u'CHAMPION', u'REIMS RUE COURCELLES'], ['REIMS', '51100', 'MARNE', '51454']],
            [[u'LECLERC', u'USSEL'], ['USSEL', '19200', 'CORREZE', '19275']],
            [[u'CHAMPION', u'NEVERS AV GAL DE GAULLE'], ['NEVERS', '58000', 'NIEVRE', '58194']],
            [[u'CARREFOUR MARKET', u"ST ANDRE DE L'EURE"], ['ST ANDRE DE L EURE', '27220', 'EURE', '27507']],
            [[u'SUPER U', u'BASSENS'], ['BASSENS', '33530', 'GIRONDE', '33032']],
            [[u'CARREFOUR MARKET', u'BEAUMONT'], ['BEAUMONT', '63110', 'PUY DE DOME', '63032']],
            [[u'GEANT CASINO', u"CHATEAU D'OLONNE"], ['CHATEAU D OLONNE', '85180', 'VENDEE', '85060']],
            [[u'LECLERC', u'ST NAZAIRE'], ['ST NAZAIRE', '44600', 'LOIRE ATLANTIQUE', '44184']],
            [[u'CARREFOUR', u'LYON 3 PART DIEU'], ['LYON 3EME ARRONDISSEMENT', '69003', 'RHONE', '69383']],
            [[u'CHAMPION', u"ST JEAN D'AULPS"], ['ST JEAN D AULPS', '74430', 'HAUTE SAVOIE', '74238']],
            [[u'INTERMARCHE', u'LES MEES'], ['LES MEES', '04190', 'ALPES DE HAUTE PROVENCE', '04116']],
            [[u'CARREFOUR MARKET', u'MOUTIERS TARENTAISE'], ['MOUTIERS', '73600', 'SAVOIE', '73181']],
            [[u'LECLERC', u'BERGERAC AV PASTEUR'], ['BERGERAC', '24100', 'DORDOGNE', '24037']],
            [[u'LECLERC', u'ST LOUP'], ['ST LOUP', '69490', 'RHONE', '69223']],
            [[u'CHAMPION', u'REIMS RUE NEUFCHATEL'], ['REIMS', '51100', 'MARNE', '51454']],
            [[u'GEANT CASINO', u'MARSEILLE 9 TASSIGNY'], ['MARSEILLE 9EME ARRONDISSE', '13009', 'BOUCHES DU RHONE', '13209']],
            [[u'CHAMPION', u'SOLESMES'], ['SOLESMES', '59730', 'NORD', '59571']],
            [[u'SUPER U', u"LE LION D'ANGERS"], ['LE LION D ANGERS', '49220', 'MAINE ET LOIRE', '49176']],
            [[u'CHAMPION', u"BLAINVILLE SUR L'EAU"], ['BLAINVILLE SUR L EAU', '54360', 'MEURTHE ET MOSELLE', '54076']],
            [[u'INTERMARCHE', u'MARSEILLE 13 CROIX ROUGE'], ['MARSEILLE 13EME ARRONDISS', '13013', 'BOUCHES DU RHONE', '13213']],
            [[u'CHAMPION', u"ST MARTIAL D'ALBAREDE"], ['ST MARTIAL D ALBAREDE', '24160', 'DORDOGNE', '24448']],
            [[u'SUPER U', u'CHATENOIS'], ['CHATENOIS', '88170', 'VOSGES', '88095']],
            [[u'SUPER U', u"PARIGNE L'EVEQUE"], ['PARIGNE L EVEQUE', '72250', 'SARTHE', '72231']],
            [[u'GEANT CASINO', u"ST MARTIN D'HERES"], ['ST MARTIN D HERES', '38400', 'ISERE', '38421']],
            [[u'GEANT CASINO', u'VALENCE CC VALENCE2'], ['VALENCE', '26000', 'DROME', '26362']],
            [[u'CARREFOUR MARKET', u'MORANGIS'], ['MORANGIS', '91420', 'ESSONNE', '91432']],
            [[u'INTERMARCHE', u'RENNES CC Italie'], ['RENNES', '35000', 'ILLE ET VILAINE', '35238']],
            [[u'INTERMARCHE', u'MERVILLE'], ['MERVILLE', '59660', 'NORD', '59400']],
            [[u'LECLERC', u'CHAMBRY'], ['CHAMBRY', '02000', 'AISNE', '02157']],
            [[u'CARREFOUR MARKET', u'MAURIAC'], ['MAURIAC', '15200', 'CANTAL', '15120']],
            [[u'CHAMPION', u'MAULEON SOULE'], ['MAULEON LICHARRE', '64130', 'PYRENEES ATLANTIQUES', '64371']], # check
            [[u'GEANT CASINO', u'MARSEILLE 12 CAILLOLS'], ['MARSEILLE 12EME ARRONDISS', '13012', 'BOUCHES DU RHONE', '13212']],
            [[u'LECLERC', u'CERNAY'], ['CERNAY', '68700', 'HAUT RHIN', '68063']], # seventh period
            [[u'LECLERC', u'LANNION ST ELIVET'], ['LANNION', '22300', "COTES D'ARMOR", '22113']],
            [[u'LECLERC', u'MARENNES'], ['MARENNES', '17320', 'CHARENTE MARITIME', '17219']],
            [[u'LECLERC', u"GONFREVILLE L'ORCHER"], ['GONFREVILLE L ORCHER', '76700', 'SEINE MARITIME', '76305']],
            [[u'CARREFOUR', u"L'ISLE ADAM"], ['L ISLE ADAM', '95290', "VAL D'OISE", '95313']],
            [[u'CARREFOUR MARKET', u'LIMOGES RUE PAGUENAUD'], ['LIMOGES', '87000', 'HAUTE VIENNE', '87085']],
            [[u'GEANT CASINO', u'MONTPELLIER AV DE LODEVE'], ['MONTPELLIER', '34000', 'HERAULT', '34172']],
            [[u'LECLERC', u'MARGON'], ['MARGON', '28400', 'EURE ET LOIR', '28236']],
            [[u'CARREFOUR MARKET', u'VERNOUILLET'], ['VERNOUILLET', '78540', 'YVELINES', '78643']],
            [[u'CARREFOUR MARKET', u'ST GRATIEN'], ['ST GRATIEN', '95210', "VAL D'OISE", '95555']],
            [[u'GEANT CASINO', u'CHAMPNIERS'], ['CHAMPNIERS', '16430', 'CHARENTE', '16078']],
            [[u'CARREFOUR', u'VERNON'], ['VERNON', '27200', 'EURE', '27681']],
            [[u'LECLERC', u'LANNION RTE PERROS GUIRREC'], ['LANNION', '22300', "COTES D'ARMOR", '22113']],
            [[u'CARREFOUR MARKET', u'ST CYPRIEN'], ['ST CYPRIEN', '24220', 'DORDOGNE', '24396']],
            [[u'CARREFOUR MARKET', u'PLOUHINEC'], ['PLOUHINEC', '29780', 'FINISTERE', '29197']],
            [[u'CARREFOUR MARKET', u"VILLENEUVE D'ASCQ"], ['VILLENEUVE D ASCQ', '59650', 'NORD', '59009']],
            [[u'AUCHAN', u'TOULOUSE CC OCCITANIA'], ['TOULOUSE', '31000', 'HAUTE GARONNE', '31555']],
            [[u'SUPER U', u'LA ROCHELLE'], ['LA ROCHELLE', '17000', 'CHARENTE MARITIME', '17300']],
            [[u'CARREFOUR MARKET', u'BOISMONT'], ['BOISMONT', '80230', 'SOMME', '80110']],
            [[u'LECLERC', u'ST PRIX'], ['ST PRIX', '95390', "VAL D'OISE", '95574']],
            [[u'INTERMARCHE', u'SERVON'], ['SERVON', '77170', 'SEINE ET MARNE', '77450']],
            [[u'CARREFOUR', u'CHATEAUROUX'], ['CHATEAUROUX', '36000', 'INDRE', '36044']],
            [[u'INTERMARCHE', u'PONS'], ['PONS', '17800', 'CHARENTE MARITIME', '17283']],
            [[u'INTERMARCHE', u'BORDEAUX'], ['BORDEAUX', '33200', 'GIRONDE', '33063']],
            [[u'CARREFOUR MARKET', u'LA ROCHETTE'], ['LA ROCHETTE', '73110', 'SAVOIE', '73215']],
            [[u'LECLERC', u'VITROLLES'], ['VITROLLES', '13127', 'BOUCHES DU RHONE', '13117']],
            [[u'LECLERC', u'ANGERVILLE'], ['ANGERVILLE', '91670', 'ESSONNE', '91016']],
            [[u'GEANT CASINO', u'ANGERS CC ESPACE ANJOU'], ['ANGERS', '49000', 'MAINE ET LOIRE', '49007']],
            [[u'SUPER U', u'ROMANS'], ['ROMANS SUR ISERE', '26100', 'DROME', '26281']],
            [[u'SUPER U', u'LYON 03 RUE F.MISTRAL'], ['LYON 3EME ARRONDISSEMENT', '69003', 'RHONE', '69383']],
            [[u'LECLERC', u'FRANCONVILLE'], ['FRANCONVILLE', '95130', "VAL D'OISE", '95252']],
            [[u'CARREFOUR', u'MEYLAN'], ['MEYLAN', '38240', 'ISERE', '38229']],
            [[u'CARREFOUR MARKET', u'CHARLEVAL'], ['CHARLEVAL', '27380', 'EURE', '27151']],
            [[u'LECLERC', u'PONS'], ['PONS', '17800', 'CHARENTE MARITIME', '17283']],
            [[u'CARREFOUR MARKET', u'CHATILLON'], ['CHATILLON', '92320', 'HAUTS DE SEINE', '92020']],
            [[u'INTERMARCHE', u"DIJON BLD DE L'UNIVERSITE"], ['DIJON', '21000', "COTE D'OR", '21231']],
            [[u'CARREFOUR', u'BEAUCAIRE'], ['BEAUCAIRE', '30300', 'GARD', '30032']],
            [[u'INTERMARCHE', u"ST CHELY D'APCHER"], ['ST CHELY D APCHER', '48200', 'LOZERE', '48140']],
            [[u'GEANT CASINO', u'MONTPELLIER CC ODYSSEUM'], ['MONTPELLIER', '34090', 'HERAULT', '34172']],
            [[u'CARREFOUR MARKET', u'BELLEY'], ['BELLEY', '01300', 'AIN', '01034']],
            [[u'CARREFOUR', u'SENS RTE DE VOULX'], ['SENS', '89100', 'YONNE', '89387']],
            [[u'GEANT CASINO', u'MARSEILLE 12 CC CAILLOLS'], ['MARSEILLE 12EME ARRONDISS', '13012', 'BOUCHES DU RHONE', '13212']],
            [[u'CARREFOUR MARKET', u'BONNEVILLE'], ['BONNEVILLE', '74130', 'HAUTE SAVOIE', '74042']],
            [[u'GEANT CASINO', u'LUCE'], ['LUCE', '28110', 'EURE ET LOIR', '28218']],
            [[u'LECLERC', u"CIVRIEUX D'AZERGUES"], ['CIVRIEUX D AZERGUES', '69380', 'RHONE', '69059']],
            [[u'GEANT CASINO', u'MARSEILLE 09 AV TASSIGNY'], ['MARSEILLE 9EME ARRONDISSE', '13009', 'BOUCHES DU RHONE', '13209']],
            [[u'CARREFOUR MARKET', u'SALLES'], ['SALLES', '33770', 'GIRONDE', '33498']],
            [[u'CARREFOUR MARKET', u'LIMOGES RUE F. PERRIN'], ['LIMOGES', '87000', 'HAUTE VIENNE', '87085']],
            [[u'INTERMARCHE', u'DIJON RUE DES 3 FORGERONS'], ['DIJON', '21000', "COTE D'OR", '21231']],
            [[u'GEANT CASINO', u'PARIS 13 CC MASSENA'], ['PARIS 13EME ARRONDISSEMENT', '75013', 'PARIS', '75113']],
            [[u'INTERMARCHE', u'ST ETIENNE RUE N. CHAIZE'], ['ST ETIENNE', '42100', 'LOIRE', '42218']],
            [[u'INTERMARCHE', u'ST ETIENNE RICHELANDIERE'], ['ST ETIENNE', '42100', 'LOIRE', '42218']],
            [[u'AUCHAN', u'MARSEILLE PONT DE VIVAUX'], ['MARSEILLE 10EME ARRONDISS', '13010', 'BOUCHES DU RHONE', '13210']],
            [[u'LECLERC', u'CHAMPFLEURY'], ['CHAMPFLEURY', '51500', 'MARNE', '51115']],
            [[u'GEANT CASINO', u'CARCASSONNE AV S.FRANCAIS'], ['CARCASSONNE', '11000', 'AUDE', '11069']],
            [[u'CARREFOUR', u'SENS RTE DE MAILLOT'], ['SENS', '89100', 'YONNE', '89387']],
            [[u'GEANT CASINO', u'MARSEILLE 08 CC MICHELET'], ['MARSEILLE 8EME ARRONDISSE', '13008', 'BOUCHES DU RHONE', '13208']],
            [[u'CARREFOUR', u'BEAUNE'], ['BEAUNE', '21200', "COTE D'OR", '21054']],
            [[u'CARREFOUR MARKET', u'MERIGNAC'], ['MERIGNAC', '33700', 'GIRONDE', '33281']], # both carrefour & c. market in M
            [[u'CARREFOUR MARKET', u'CIVRAY'], ['CIVRAY', '86400', 'VIENNE', '86078']], # eighth period
            [[u'CARREFOUR', u'LA ROCHE S. YON RTE DE NANTES'], ['LA ROCHE SUR YON', '85000', 'VENDEE', '85191']],
            [[u'LECLERC', u'LANNION ROUTE DE GUIGAMP'], ['LANNION', '22300', "COTES D'ARMOR", '22113']],
            [[u'CARREFOUR', u'LYON C C PART DIEU'], ['LYON 3EME ARRONDISSEMENT', '69003', 'RHONE', '69383']],
            [[u'CARREFOUR MARKET', u"VILLENEUVE L'ARCHEVEQUE"], ['VILLENEUVE L ARCHEVEQUE', '89190', 'YONNE', '89461']],
            [[u'CARREFOUR', u'VITROLLES'], ['VITROLLES', '13127', 'BOUCHES DU RHONE', '13117']],
            [[u'INTERMARCHE', u'ORGEVAL'], ['ORGEVAL', '78630', 'YVELINES', '78466']],
            [[u'CARREFOUR MARKET', u"BLAINVILLE SUR L'EAU"], ['BLAINVILLE SUR L EAU', '54360', 'MEURTHE ET MOSELLE', '54076']],
            [[u'CARREFOUR', u'SAINT MARTIN AU LAERT'], ['ST MARTIN AU LAERT', '62500', 'PAS DE CALAIS', '62757']],
            [[u'CARREFOUR', u'LESPAR MEDOC'], ['LESPARRE MEDOC', '33340', 'GIRONDE', '33240']],
            [[u'CARREFOUR MARKET', u'PARIS 13 C C Italie 2'], ['PARIS 13EME ARRONDISSEMENT', '75013', 'PARIS', '75113']],
            [[u'LECLERC', u'MASSY'], ['MASSY', '91300', 'ESSONNE', '91377']],
            [[u'CARREFOUR MARKET', u'BELLEGARDE'], ['BELLEGARDE', '30127', 'GARD', '30034']], # not 100% sure
            [[u'INTERMARCHE', u'MALICORNE'], ['MALICORNE', '03600', 'ALLIER', '03159']],
            [[u'CARREFOUR', u'LA ROCHE S. YON RTE DES SABLES'], ['LA ROCHE SUR YON', '85000', 'VENDEE', '85191']],
            [[u'INTERMARCHE', u'RUFFEC'], ['RUFFEC', '16700', 'CHARENTE', '16292']],
            [[u'LECLERC', u'LANNION ROUTE PERROS GUIREC'], ['LANNION', '22300', "COTES D'ARMOR", '22113']],
            [[u'SUPER U', u"ST SEURIN SUR L'ISLE"], ['ST SEURIN SUR L ISLE', '33660', 'GIRONDE', '33478']],
            [[u'CARREFOUR', u'SENS ROUTE DE MAILLOT'], ['SENS', '89100', 'YONNE', '89387']],
            [[u'GEANT CASINO', u'MARSEILLE C C CAILLOLS'], ['MARSEILLE 12EME ARRONDISS', '13012', 'BOUCHES DU RHONE', '13212']],
            [[u'GEANT CASINO', u'MARSEILLE LA VALENTINE'], ['MARSEILLE 11EME ARRONDISS', '13011', 'BOUCHES DU RHONE', '13211']],
            [[u'LECLERC', u'GENAY'], ['GENAY', '69730', 'RHONE', '69278']],
            [[u'SUPER U', u"L'HERMITAGE"], ['L HERMITAGE', '35590', 'ILLE ET VILAINE', '35131']],
            [[u'CARREFOUR', u'MOULINS'], ['MOULINS', '03000', 'ALLIER', '03190']],
            [[u'CARREFOUR', u'CHELLES'], ['CHELLES', '77500', 'SEINE ET MARNE', '77108']],
            [[u'CARREFOUR', u'SENS ROUTE DE VOULX'], ['SENS', '89100', 'YONNE', '89387']],
            [[u'CARREFOUR MARKET', u'PARIS 18 AVE DE ST OUEN'], ['PARIS 18EME ARRONDISSEMENT', '75018', 'PARIS', '75118']],
            [[u'SUPER U', u'MONDEVILLE'], ['MONDEVILLE', '14120', 'CALVADOS', '14437']],
            [[u'CARREFOUR', u'PARIS 16 AVENUE DU GAL SARRAIL'], ['PARIS 16EME ARRONDISSEMENT', '75016', 'PARIS', '75116']],
            [[u'CARREFOUR', u'MONTESSON'], ['MONTESSON', '78360', 'YVELINES', '78418']],
            [[u'CARREFOUR MARKET', u'SAUSSAY'], ['SAUSSAY', '28260', 'EURE ET LOIR', '28371']],
            [[u'GEANT CASINO', u'VALENCE C C VALENCE 2'], ['VALENCE', '26000', 'DROME', '26362']],
            [[u'CARREFOUR MARKET', u'ST MAURICE'], ['ST MAURICE', '94410', 'VAL DE MARNE', '94069']],
            [[u'CARREFOUR MARKET', u'PARIS 18 CLIGNANCOURT'], ['PARIS 18EME ARRONDISSEMENT', '75018', 'PARIS', '75118']],
            [[u'GEANT CASINO', u'MARSEILLE AVE DE MARZAGUES'], ['MARSEILLE 8EME ARRONDISSE', '13008', 'BOUCHES DU RHONE', '13208']],
            [[u'AUCHAN', u'PERONNE'], ['PERONNE', '80200', 'SOMME', '80620']],
            [[u'CARREFOUR', u'FRANCHEVILLE'], ['FRANCHEVILLE', '69340', 'RHONE', '69089']],
            [[u'INTERMARCHE', u'LAROQUE'], ['LAROQUE', '34190', 'HERAULT', '34128']],
            [[u'INTERMARCHE', u"L'ISLE JOURDAIN"], ['L ISLE JOURDAIN', '86150', 'VIENNE', '86112']],
            [[u'CARREFOUR MARKET', u'GRANDVILLIERS'], ['GRANDVILLIERS', '60210', 'OISE', '60286']],
            [[u'CARREFOUR MARKET', u'MONTMIRAIL'], ['MONTMIRAIL', '51210', 'MARNE', '51380']],
            [[u'LECLERC', u'LES ANGLES'], ['LES ANGLES', '30133', 'GARD', '30011']],
            [[u'GEANT CASINO', u'CARCASSONNE AVE DU SOUVENIR'], ['CARCASSONNE', '11000', 'AUDE', '11069']],
            [[u'CARREFOUR MARKET', u'MONTBRISON'], ['MONTBRISON', '42600', 'LOIRE', '42147']],
            [[u'CARREFOUR MARKET', u'ST MARTIN VIEUX BELLEME'], ['ST MARTIN DU VIEUX BELLEM', '61130', 'ORNE', '61426']],
            [[u'LECLERC', u'ROQUES'], ['ROQUES', '31120', 'HAUTE GARONNE', '31458']],
            [[u'INTERMARCHE', u"VABRES L'ABBAYE"], ['VABRES L ABBAYE', '12400', 'AVEYRON', '12286']],
            [[u'GEANT CASINO', u'CARCASSONNE C C SALVAZA'], ['CARCASSONNE', '11000', 'AUDE', '11069']],
            [[u'GEANT CASINO', u'MONTPELLIER AVE DE LODEVE'], ['MONTPELLIER', '34000', 'HERAULT', '34172']],
            [[u'GEANT CASINO', u'MONTPELLIER C C ODYSSEUM'], ['MONTPELLIER', '34090', 'HERAULT', '34172']],
            [[u'CARREFOUR MARKET', u'ST ANDRE DE LA ROCHE'], ['ST ANDRE', '06730', 'ALPES MARITIMES', '06114']],# ninth period
            [[u'SUPER U', u"ST BARTHELEMY D'ANJOU"], ['ST BARTHELEMY D ANJOU', '49124', 'MAINE ET LOIRE', '49267']],
            [[u'INTERMARCHE', u"L'UNION"], ['L UNION', '31240', 'HAUTE GARONNE', '31561']],
            [[u'CARREFOUR MARKET', u'BONNEVAL'], ['BONNEVAL', '28800', 'EURE ET LOIR', '28051']],
            [[u'CARREFOUR', u'MARSEILLE CC LE MERLAN'], ['MARSEILLE 14EME ARRONDISS', '13014', 'BOUCHES DU RHONE', '13214']],
            [[u'CARREFOUR MARKET', u'SAINT DENIS LES BOURGS'], ['ST DENIS LES BOURG', '01000', 'AIN', '01344']],
            [[u'CARREFOUR MARKET', u'PARIGNY'], ['PARIGNY', '50600', 'MANCHE', '50391']],
            [[u'CARREFOUR', u'NICE CC NTL'], ['NICE', '06300', 'ALPES MARITIMES', '06088']],
            [[u'CARREFOUR', u"L'ISLE D'ABEAU"], ['L ISLE D ABEAU', '38080', 'ISERE', '38193']],
            [[u'CARREFOUR MARKET', u'BELLEVILLE'], ['BELLEVILLE SUR SAONE', '69220', 'RHONE', '69019']], # not 100%
            [[u'AUCHAN', u'SAINT JEAN LA RUELLE'], ['ST JEAN DE LA RUELLE', '45140', 'LOIRET', '45285']],
            [[u'GEANT CASINO', u'MARSEILLE CC VALENTINE'], ['ST JEAN DE LA RUELLE', '45140', 'LOIRET', '45285']],
            [[u'CARREFOUR', u'NICE - N202'],  ['NICE', '06200', 'ALPES MARITIMES', '06088']],
            [[u'CARREFOUR MARKET', u'CHATEAURENARD'], ['CHATEAURENARD', '13160', 'BOUCHES DU RHONE', '13027']],
            [[u'CARREFOUR', u'ANGERS CC ST SERGE'], ['ANGERS', '49100', 'MAINE ET LOIRE', '49007']],
            [[u'LECLERC', u'SAINTS GEOSMES'], ['LANGRES', '52200', 'HAUTE MARNE', '52269']],
            [[u'GEANT CASINO', u'CARCASSONNE SOUV.FRANCAIS'], ['CARCASSONNE', '11000', 'AUDE', '11069']],
            [[u'GEANT CASINO', u'MARSEILLE CC MICHELET'], ['MARSEILLE 8EME ARRONDISSE', '13008', 'BOUCHES DU RHONE', '13208']],
            [[u'CARREFOUR', u'ANGERS CC GD MAINE'], ['ANGERS', '49000', 'MAINE ET LOIRE', '49007']],
            [[u'HYPER U', u'LES ARCS'], ['LES ARCS', '83460', 'VAR', '83004']],
            [[u'SUPER U', u'SERQUEUX'], ['SERQUEUX', '76440', 'SEINE MARITIME', '76672']],
            [[u'LECLERC', u'MONTDIDIER'], ['MONTDIDIER', '80500', 'SOMME', '80561']],
            [[u'CARREFOUR MARKET', u'PARIS 14 GAL LECLERC'], ['PARIS 14EME ARRONDISSEMENT', '75014', 'PARIS', '75114']],
            [[u'HYPER U', u'VITRE'],  ['VITRE', '35500', 'ILLE ET VILAINE', '35360']],
            [[u'LECLERC', u'MARLY'], ['MARLY', '57155', 'MOSELLE', '57447']],
            [[u'INTERMARCHE', u"ST PIERRE D'ALBIGNY"], ['ST PIERRE D ALBIGNY', '73250', 'SAVOIE', '73270']],
            [[u'LECLERC', u'MONTLUCON AV PDT AURIOL'], ['MONTLUCON', '03100', 'ALLIER', '03185']],
            [[u'CARREFOUR MARKET', u'DOMFRONT'], ['DOMFRONT', '61700', 'ORNE', '61145']],
            [[u'CARREFOUR', u"CONDE SUR L'ESCAUT"], ['CONDE SUR L ESCAUT', '59163', 'NORD', '59153']],
            [[u'SUPER U', u'LYON 03 F.MISTRAL'], ['LYON 3EME ARRONDISSEMENT', '69003', 'RHONE', '69383']],
            [[u'LECLERC', u'LURE'], ['LURE', '70200', 'HAUTE SAONE', '70310']],
            [[u'CARREFOUR', u'MARSEILLE CC GD LITTORAL'], ['MARSEILLE 15EME ARRONDISS', '13015', 'BOUCHES DU RHONE', '13215']],
            [[u'SUPER U', u"L'HORME"], ['L HORME', '42152', 'LOIRE', '42110']],
            [[u'LECLERC', u"ST MARTIN D'HERES"], ['ST MARTIN D HERES', '38400', 'ISERE', '38421']],
            [[u'SUPER U', u'LA POMMERAYE'], ['LA POMMERAYE', '49620', 'MAINE ET LOIRE', '49244']],
            [[u'LECLERC', u'MONTLUCON AVE P.VILON'], ['MONTLUCON', '03100', 'ALLIER', '03185']],
            [[u'CARREFOUR MARKET', u'NEVERS COLBERT'], ['NEVERS', '58000', 'NIEVRE', '58194']],
            [[u'INTERMARCHE', u'SAINT JEAN DE MUZOLS'], ['ST JEAN DE MUZOLS', '07300', 'ARDECHE', '07245']],
            [[u'GEANT CASINO', u'MONTPELLIER AVE LODEVE'], ['MONTPELLIER', '34000', 'HERAULT', '34172']], # 34070?
            [[u'CARREFOUR MARKET', u'MIRIBEL'],  ['MIRIBEL', '01700', 'AIN', '01249']],
            [[u'CARREFOUR MARKET', u'MARSEILLAN'], ['MARSEILLAN', '34340', 'HERAULT', '34150']],
            [[u'SUPER U', u'PARIS 17 AVE DE CLICHY'], ['PARIS 17EME ARRONDISSEMENT', '75017', 'PARIS', '75117']],
            [[u'GEANT CASINO', u'MONTPELLIER AVE ARGELLIERS'], ['MONTPELLIER', '34000', 'HERAULT', '34172']],
            [[u'CARREFOUR', u"L'HAY LES ROSES"], ['L HAY LES ROSES', '94240', 'VAL DE MARNE', '94038']],
            [[u'LECLERC', u'BUXEROLLES'], ['BUXEROLLES', '86180', 'VIENNE', '86041']],
            [[u'SUPER U', u'PASSY'], ['PASSY', '74190', 'HAUTE SAVOIE', '74208']],
            [[u'LECLERC', u'LYON 09 CHAMPVERT'], ['LYON 9EME ARRONDISSEMENT', '69009', 'RHONE', '69389']],
            [[u'GEANT CASINO', u'MARSEILLE CC CAILLOLS'], ['MARSEILLE 12EME ARRONDISS', '13012', 'BOUCHES DU RHONE', '13212']],
            [[u'LECLERC', u"PONT L'ABBE"], ['PONT L ABBE', '29120', 'FINISTERE', '29220']],
            [[u'LECLERC', u'NORMANVILLE'], ['NORMANVILLE', '27930', 'EURE', '27439']],
            [[u'CARREFOUR', u'ST MARTIN DES CHAMPS'], ['ST MARTIN DES CHAMPS', '50300', 'MANCHE', '50516']],
            [[u'SUPER U', u"ST BARTHELEMY D''ANJOU"], ['ST BARTHELEMY D ANJOU', '49124', 'MAINE ET LOIRE', '49267']], # tenth
            [[u'CARREFOUR MARKET', u'CLERMONT CARRE JAUDE'],  ['CLERMONT FERRAND', '63000', 'PUY DE DOME', '63113']],
            [[u'INTERMARCHE', u'LA ROCHELLE'],  ['LA ROCHELLE', '17000', 'CHARENTE MARITIME', '17300']],
            [[u'LECLERC', u'SAINT AMAND MONTROND'], ['ST AMAND MONTROND', '18200', 'CHER', '18197']],
            [[u'SUPER U', u'MARANS'], ['MARANS', '17230', 'CHARENTE MARITIME', '17218']],
            [[u'GEANT CASINO', u"VILLENAVE D''ORNON"], ['VILLENAVE D ORNON', '33140', 'GIRONDE', '33550']],
            [[u'LECLERC', u'LA ROCHELLE'],  ['LA ROCHELLE', '17000', 'CHARENTE MARITIME', '17300']],
            [[u'CORA', u"VILLENEUVE D''ASCQ"], ['VILLENEUVE D ASCQ', '59650', 'NORD', '59009']],
            [[u'SUPER U', u'MUNSTER'], ['MUNSTER', '68140', 'HAUT RHIN', '68226']],
            [[u'GEANT CASINO', u'CARCASSONNE AV DU SOUVENIR'], ['CARCASSONNE', '11000', 'AUDE', '11069']],
            [[u'CARREFOUR MARKET', u"AIRE SUR L''ADOUR"], ['AIRE SUR L ADOUR', '40800', 'LANDES', '40001']],
            [[u'AUCHAN', u'BORDEAUX QUARTIER DU LAC'], ['BORDEAUX', '33000', 'GIRONDE', '33063']],
            [[u'GEANT CASINO', u'MONTPELLIER PLACE LISBONNE'], ['MONTPELLIER', '34000', 'HERAULT', '34172']],
            [[u'CARREFOUR', u'CHALON SUR SAONE CC CHALON SUD'],  ['CHALON SUR SAONE', '71100', 'SAONE ET LOIRE', '71076']],
            [[u'CARREFOUR MARKET', u'PARIS 18'], ['PARIS 18EME ARRONDISSEMENT', '75018', 'PARIS', '75118']],
            [[u'CARREFOUR MARKET', u'PARIS 15'], ['PARIS 15EME ARRONDISSEMENT', '75015', 'PARIS', '75115']],
            [[u'INTERMARCHE', u'SAINT LAURENT DU VAR'], ['ST LAURENT DU VAR', '06700', 'ALPES MARITIMES', '06123']],
            [[u'INTERMARCHE', u'LE MANS AV BOLLEE'], ['LE MANS', '72000', 'SARTHE', '72181']],
            [[u'GEANT CASINO', u'CARCASSONNE ZI LA BOURIETTE'], ['CARCASSONNE', '11000', 'AUDE', '11069']],
            [[u'CARREFOUR MARKET', u'CLERMONT FERRAND ALYRE'], ['CLERMONT FERRAND', '63000', 'PUY DE DOME', '63113']],
            [[u'CARREFOUR MARKET', u'ST PRIEST'], ['ST PRIEST', '69800', 'RHONE', '69290']],
            [[u'LECLERC', u'ALLONNES'], ['ALLONNES', '72700', 'SARTHE', '72003']],
            [[u'INTERMARCHE', u'SAINT NIZIER SOUS CHARLIEU'], ['ST NIZIER SOUS CHARLIEU', '42190', 'LOIRE', '42267']],
            [[u'INTERMARCHE', u'AVRILLE'], ['AVRILLE', '49240', 'MAINE ET LOIRE', '49015']],
            [[u'HYPER U', u'FALAISE'], ['FALAISE', '14700', 'CALVADOS', '14258']],
            [[u'GEANT CASINO', u'BAZEILLE'], ['BAZEILLES', '08140', 'ARDENNES', '08053']],
            [[u'CARREFOUR MARKET', u'BORDEAUX'], ['BORDEAUX', '33000', 'GIRONDE', '33063']],
            [[u'CARREFOUR MARKET', u'ST OUEN'], ['PARIS 18EME ARRONDISSEMENT', '75018', 'PARIS', '75118']], # not 100%
            [[u'GEANT CASINO', u'MARSEILLE 09'], ['MARSEILLE 9EME ARRONDISSE', '13009', 'BOUCHES DU RHONE', '13209']],
            [[u'GEANT CASINO', u'MARSEILLE 08'], ['MARSEILLE 8EME ARRONDISSE', '13008', 'BOUCHES DU RHONE', '13208']],
            [[u'CARREFOUR', u'CREYSSE'], ['CREYSSE', '24100', 'DORDOGNE', '24145']], # => intermarche
            [[u'SUPER U', u"ST SYLVAIN D''ANJOU"], ['ST SYLVAIN D ANJOU', '49480', 'MAINE ET LOIRE', '49323']],
            [[u'INTERMARCHE', u"L''ISLE SUR LE DOUBS"], ['L ISLE SUR LE DOUBS', '25250', 'DOUBS', '25315']],
            [[u'AUCHAN', u'PLAN DE GRASSE'], ['GRASSE', '06130', 'ALPES MARITIMES', '06069']],
            [[u'CARREFOUR', u'FLINS'], ['FLINS SUR SEINE', '78410', 'YVELINES', '78238']],
            [[u'INTERMARCHE', u'AULNAY'], ['AULNAY', '17470', 'CHARENTE MARITIME', '17024']], # not 100%
            [[u'INTERMARCHE', u'MERIGNAC'], ['MERIGNAC', '33700', 'GIRONDE', '33281']],
            [[u'INTERMARCHE', u'MONTIGNAC'], ['MONTIGNAC', '24290', 'DORDOGNE', '24291']],
            [[u'LECLERC', u'ACHERES'], ['ACHERES', '78260', 'YVELINES', '78005']],
            [[u'GEANT CASINO', u'MONTPELLIER MAS ARGELLIERS'], ['MONTPELLIER', '34000', 'HERAULT', '34172']],
            [[u'LECLERC', u'SELESTAT ALLEE WESTRICH'], ['SELESTAT', '67600', 'BAS RHIN', '67462']],
            [[u'LECLERC', u'LAVAL'], ['LAVAL', '53000', 'MAYENNE', '53130']],
            [[u'CARREFOUR MARKET', u'MOUTIERS'], ['MOUTIERS', '73600', 'SAVOIE', '73181']],
            [[u'LECLERC', u"PONT L''ABBE"], ['PONT L ABBE', '29120', 'FINISTERE', '29220']],
            [[u'CARREFOUR', u'CHALON SUR SAONE AV DE PARIS'], ['CHALON SUR SAONE', '71100', 'SAONE ET LOIRE', '71076']],
            [[u'SUPER U', u'LE VIGAN'], ['LE VIGAN', '30120', 'GARD', '30350']],
            [[u'INTERMARCHE', u"ST CHELY D''APCHER"], ['ST CHELY D APCHER', '48200', 'LOZERE', '48140']],
            [[u'CARREFOUR MARKET', u'CLERMONT FERRAND JAUDE'], ['CLERMONT FERRAND', '63000', 'PUY DE DOME', '63113']],
            [[u'INTERMARCHE', u'ST GILLES'], ['ST GILLES', '30800', 'GARD', '30258']],
            [[u'SUPER U', u"ST DONAT SUR L''HERBASSE"], ['ST DONAT SUR L HERBASSE', '26260', 'DROME', '26301']],
            [[u'CARREFOUR', u'NIMES CC NIMES SUD'], ['NIMES', '30000', 'GARD', '30189']],
            [[u'AUCHAN', u"VILLENEUVE D''ASCQ"], ['VILLENEUVE D ASCQ', '59650', 'NORD', '59009']],
            [[u'LECLERC', u'SELESTAT ALLEE DE LA LOHMUHLE'], ['SELESTAT', '67600', 'BAS RHIN', '67462']],
            [[u'CARREFOUR MARKET', u'LAVAL'], ['LAVAL', '53000', 'MAYENNE', '53130']],
            [[u'E. LECLERC', u'VALENCE'], ['VALENCE', '26000', 'DROME', '26362']], # eleventh period
            [[u'E. LECLERC', u'ST MAUR'], ['ST MAUR', '36250', 'INDRE', '36202']],
            [[u'CARREFOUR', u'PARIS 16'], ['PARIS 16EME ARRONDISSEMENT', '75016', 'PARIS', '75116']],
            [[u'INTERMARCHE', u'CAMON'], ['CAMON', '80450', 'SOMME', '80164']],
            [[u'SUPER U', u'MARSEILLE ST PIERRE'], ['MARSEILLE 5EME ARRONDISSE', '13005', 'BOUCHES DU RHONE', '13205']],
            [[u'INTERMARCHE', u'ST APOLLINAIRE'], ['ST APOLLINAIRE', '21850', "COTE D'OR", '21540']],
            [[u'E. LECLERC', u'ROQUES'], ['ROQUES', '31120', 'HAUTE GARONNE', '31458']],
            [[u'INTERMARCHE', u'LE PONT DE BEAUVOISIN'], ['LE PONT DE BEAUVOISIN', '73330', 'SAVOIE', '73204']],
            [[u'SUPER U', u'IDRON'], ['IDRON-OUSSE-SENDETS', '64320', 'PYRENEES ATLANTIQUES', '64269']],
            [[u'SUPER U', u'CUIRASSE BREST'], ['BREST', '29200', 'FINISTERE', '29019']],
            [[u'GEANT CASINO', u"MONTPELLIER D''ARGELLIERS"], ['MONTPELLIER', '34000', 'HERAULT', '34172']],
            [[u'E. LECLERC', u'BEAUNE'], ['BEAUNE', '21200', "COTE D'OR", '21054']],
            [[u'E. LECLERC', u'GENAY'], ['GENAY', '69730', 'RHONE', '69278']],
            [[u'CARREFOUR MARKET', u'LIMOGES PAGUENAUD'], ['LIMOGES', '87000', 'HAUTE VIENNE', '87085']],
            [[u'E. LECLERC', u'LES ANGLES'], ['LES ANGLES', '30133', 'GARD', '30011']],
            [[u'GEANT CASINO', u'CARCASSONNE SALVAZA'], ['CARCASSONNE', '11000', 'AUDE', '11069']],
            [[u'E. LECLERC', u'MONTDIDIER'], ['MONTDIDIER', '80500', 'SOMME', '80561']],
            [[u'E. LECLERC', u'BUXEROLLES'], ['BUXEROLLES', '86180', 'VIENNE', '86041']],
            [[u'AUCHAN', u'BORDEAUX MERIADECK'], ['BORDEAUX', '33000', 'GIRONDE', '33063']],
            [[u'E. LECLERC', u'ST RAPHAEL'], ['ST RAPHAEL', '83700', 'VAR', '83118']],
            [[u'CARREFOUR MARKET', u'THONON LES BAINS LANDAZ'], ['THONON LES BAINS', '74200', 'HAUTE SAVOIE', '74281']],
            [[u'E. LECLERC', u'VITROLLES'], ['VITROLLES', '13127', 'BOUCHES DU RHONE', '13117']],
            [[u'INTERMARCHE', u'MACON PERCEVAL'], ['MACON', '71000', 'SAONE ET LOIRE', '71270']],
            [[u'GEANT CASINO', u'ANGERS ESPACE ANJOU'], ['ANGERS', '49000', 'MAINE ET LOIRE', '49007']],
            [[u'INTERMARCHE', u'AUREILHAN'], ['AUREILHAN', '65800', 'HAUTES PYRENEES', '65047']],
            [[u'CARREFOUR MARKET', u'VIRE'], ['VIRE', '14500', 'CALVADOS', '14762']],
            [[u'CARREFOUR', u'FLERS'], ['FLERS', '61100', 'ORNE', '61169']], # could be wrong
            [[u'INTERMARCHE', u'POITIERS DEMIE LUNE'], ['POITIERS', '86000', 'VIENNE', '86194']],
            [[u'INTERMARCHE', u'JUILLIAN'], ['JUILLAN', '65290', 'HAUTES PYRENEES', '65235']],
            [[u'CARREFOUR MARKET', u'GOURDON'], ['GOURDON', '46300', 'LOT', '46127']],
            [[u'CARREFOUR MARKET', u'THONON LES BAINS FERRY'], ['THONON LES BAINS', '74200', 'HAUTE SAVOIE', '74281']],
            [[u'E. LECLERC', u'CERNAY'], ['CERNAY', '68700', 'HAUT RHIN', '68063']],
            [[u'INTERMARCHE', u'CASTRES'], ['CASTRES', '81100', 'TARN', '81065']],
            [[u'CARREFOUR MARKET', u'BELLEVUE BREST'], ['BREST', '29200', 'FINISTERE', '29019']],
            [[u'CARREFOUR', u'LAVAL'], ['LAVAL', '53000', 'MAYENNE', '53130']],
            [[u'CARREFOUR MARKET', u'LIMOGES PERRIN'], ['LIMOGES', '87000', 'HAUTE VIENNE', '87085']],
            [[u'SUPER U', u'PRISSE'], ['PRISSE', '71960', 'SAONE ET LOIRE', '71360']],
            [[u'GEANT CASINO', u'GOUESNOU BREST'], ['BREST', '29200', 'FINISTERE', '29019']],
            [[u'GEANT CASINO', u'CARCASSONNE SOUV. FRAN'], ['CARCASSONNE', '11000', 'AUDE', '11069']],
            [[u'E. LECLERC', u"BOIS D''ARCY"], ['BOIS D ARCY', '78390', 'YVELINES', '78073']],
            [[u'CARREFOUR MARKET', u'LA CAPELLE'], ['LA CAPELLE', '02260', 'AISNE', '02141']],
            [[u'E. LECLERC', u'CHARMES'], ['CHARMES', '88130', 'VOSGES', '88090']],
            [[u'GEANT CASINO', u'BEZIER'], ['BEZIERS', '34500', 'HERAULT', '34032']],
            [[u'GEANT CASINO', u'MARSEILLE CAILLOLS'], ['MARSEILLE 12EME ARRONDISS', '13012', 'BOUCHES DU RHONE', '13212']],
            [[u'E. LECLERC', u'GOUESNOU BREST'], ['BREST', '29200', 'FINISTERE', '29019']],
            [[u'CARREFOUR MARKET', u'BEAUREPAIRE'], ['BEAUREPAIRE', '38270', 'ISERE', '38034']],
            [[u'GEANT CASINO', u'MASSENA PARIS'], ['PARIS 13EME ARRONDISSEMENT', '75013', 'PARIS', '75113']],
            [[u'CARREFOUR MARKET', u'BORDEAUX SAINT JEAN'], ['BORDEAUX', '33800', 'GIRONDE', '33063']],
            [[u'CARREFOUR MARKET', u'MARSEILLE L. MEISSEREL'], ['MARSEILLE 12EME ARRONDISS', '13012', 'BOUCHES DU RHONE', '13212']],
            [[u'CARREFOUR MARKET', u'SORBIERS'], ['SORBIERS', '42290', 'LOIRE', '42302']],
            [[u'GEANT CASINO', u'MONTPELLIER ODYSSEUM'], ['MONTPELLIER', '34090', 'HERAULT', '34172']],
            [[u'E. LECLERC', u'REZE OCEANE'], ['REZE', '44400', 'LOIRE ATLANTIQUE', '44143']],
            [[u'CARREFOUR MARKET', u'THEIX'], ['THEIX', '56450', 'MORBIHAN', '56251']],
            [[u'E. LECLERC', u'REZE ATOUT SUD'], ['REZE', '44400', 'LOIRE ATLANTIQUE', '44143']],
            [[u'AUCHAN', u'BORDEAUX LE LAC'], ['BORDEAUX', '33300', 'GIRONDE', '33063']],
            [[u'GEANT CASINO', u'MARSEILLE LA SABLIERE'], ['MARSEILLE 11EME ARRONDISS', '13011', 'BOUCHES DU RHONE', '13211']], # twelfth period
            [[u'INTERMARCHE', u'LOUVIGNY'], ['LOUVIGNY', '14111', 'CALVADOS', '14383']],
            [[u'CARREFOUR MARKET', u'FERRIERES EN GATINA'], ['FERRIERES EN GATINAIS', '45210', 'LOIRET', '45145']],
            [[u'CARREFOUR MARKET', u'LE FAOUET'], ['LE FAOUET', '56320', 'MORBIHAN', '56057']],
            [[u'GEANT CASINO', u'CARCASSONNE SOUVENIR Fr'], ['CARCASSONNE', '11000', 'AUDE', '11069']],
            [[u'LECLERC', u'HERICOURT'], ['HERICOURT', '70400', 'HAUTE SAONE', '70285']],
            [[u'CARREFOUR MARKET', u'ECUELLES'], ['ECUELLES', '77250', 'SEINE ET MARNE', '77166']],
            [[u'LECLERC', u'VITR\xc9'], ['VITRE', '35500', 'ILLE ET VILAINE', '35360']],
            [[u'CARREFOUR MARKET', u'ST NAZAIRE'], ['ST NAZAIRE', '44600', 'LOIRE ATLANTIQUE', '44184']], #if indeed market
            [[u'INTERMARCHE', u'BAILLEUL'], ['BAILLEUL', '59270', 'NORD', '59043']],
            [[u'LECLERC', u'ST ASTIER'], ['ST ASTIER', '24110', 'DORDOGNE', '24372']],
            [[u'CARREFOUR', u'MARSEILLE BONNEVEINE'], ['MARSEILLE 8EME ARRONDISSE', '13008', 'BOUCHES DU RHONE', '13208']],
            [[u'INTERMARCHE', u'VALENCE'], ['VALENCE', '26000', 'DROME', '26362']],
            [[u'CARREFOUR', u'MARSEILLE LE MERLAN'], ['MARSEILLE 14EME ARRONDISS', '13014', 'BOUCHES DU RHONE', '13214']],
            [[u'GEANT CASINO', u"MONTPELLIER MAS D''ARGE"], ['MONTPELLIER', '34000', 'HERAULT', '34172']],
            [[u'GEANT CASINO', u'MARSEILLE MICHELET STE'], ['MARSEILLE 8EME ARRONDISSE', '13008', 'BOUCHES DU RHONE', '13208']],
            [[u'LECLERC', u'FLERS'], ['FLERS', '61100', 'ORNE', '61169']],
            [[u'GEANT CASINO', u'MARSEILLE DE LATTRE DE'], ['MARSEILLE 9EME ARRONDISSE', '13009', 'BOUCHES DU RHONE', '13209']],
            [[u'SUPER U', u'LES ESSARTS'], ['LES ESSARTS', '85140', 'VENDEE', '85084']],
            [[u'SUPER U', u'BELZT'], ['BELZ', '56550', 'MORBIHAN', '56013']], # mispelled... hopefully right
            [[u'CARREFOUR MARKET', u'LIGNIERES'], ['LIGNIERES', '18160', 'CHER', '18127']],
            [[u'CARREFOUR MARKET', u'RAILLENCOURT STE OL'], ['RAILLENCOURT STE OLLE', '59554', 'NORD', '59488']],
            [[u'INTERMARCHE', u'SAINT-NAZAIRE'], ['ST NAZAIRE', '44600', 'LOIRE ATLANTIQUE', '44184']],
            [[u'CARREFOUR', u'NANTES LA BEAUJOIRE'], ['NANTES', '44300', 'LOIRE ATLANTIQUE', '44109']],
            [[u'HYPER U', u'LA MONTAGNE'], ['LA MONTAGNE', '44620', 'LOIRE ATLANTIQUE', '44101']],
            [[u'INTERMARCHE', u"PONT-L'\xc9V\xcaQUE"], ['PONT L EVEQUE', '14130', 'CALVADOS', '14514']],
            [[u'CARREFOUR MARKET', u'CHANGE'], ['CHANGE', '53810', 'MAYENNE', '53054']],
            [[u'INTERMARCHE', u'BEAUNE'], ['BEAUNE', '21200', "COTE D'OR", '21054']],
            [[u'CARREFOUR', u'NANTES BEAULIEU'], ['NANTES', '44200', 'LOIRE ATLANTIQUE', '44109']],
            [[u'CARREFOUR MARKET', u'AUTERIVE'], ['AUTERIVE', '31190', 'HAUTE GARONNE', '31033']], #thirteeth period
            [[u'GEANT CASINO', u'MARSEILLE DE LATTRE DE TASSIG'], ['MARSEILLE 9EME ARRONDISSE', '13009', 'BOUCHES DU RHONE', '13209']],
            [[u'INTERMARCHE SUPER', u'MONTIGNAC'], ['MONTIGNAC', '24290', 'DORDOGNE', '24291']],
            [[u'INTERMARCHE SUPER', u'LAVAL'], ['LAVAL', '53000', 'MAYENNE', '53130']],
            [[u'CARREFOUR MARKET', u'VILLERS BOCAGE'], []], # ambiguous... could be both?
            [[u'GEANT CASINO', u"MONTPELLIER MAS D''ARGELLIERS"], ['MONTPELLIER', '34000', 'HERAULT', '34172']],
            [[u'GEANT CASINO', u'MARSEILLE MICHELET STE ANNE'], ['MARSEILLE 8EME ARRONDISSE', '13008', 'BOUCHES DU RHONE', '13208']],
            [[u'CENTRE E. LECLERC', u'VITROLLES'], ['VITROLLES', '13127', 'BOUCHES DU RHONE', '13117']],
            [[u'CARREFOUR MARKET', u'SAINT-NAZAIRE'], ['ST NAZAIRE', '44600', 'LOIRE ATLANTIQUE', '44184']], #if indeed market
            [[u'CENTRE E. LECLERC', u'BAILLEUL'], ['BAILLEUL', '59270', 'NORD', '59043']],
            [[u'INTERMARCHE', u"PONT L'EVEQUE"], ['PONT L EVEQUE', '14130', 'CALVADOS', '14514']],
            [[u'INTERMARCHE SUPER', u'BORDES'], ['BORDES', '64510', 'PYRENEES ATLANTIQUES', '64138']],
            [[u'INTERMARCHE SUPER', u'ETRECHY'], ['ETRECHY', '91580', 'ESSONNE', '91226']],
            [[u'INTERMARCHE SUPER', u'BEAUNE'], ['BEAUNE', '21200', "COTE D'OR", '21054']],
            [[u'GEANT CASINO', u'CARCASSONNE SOUVENIR FRAN\xc7AIS'], ['CARCASSONNE', '11000', 'AUDE', '11069']],
            [[u'GEANT CASINO', u'CHALON SUR SAONE CC LA THALIE'], ['CHALON SUR SAONE', '71100', 'SAONE ET LOIRE', '71076']],
            [[u'CARREFOUR MARKET', u'ANNEMASSE FLORISSANT'], ['ANNEMASSE', '74100', 'HAUTE SAVOIE', '74012']],
            [[u'CENTRE E. LECLERC', u'VERDUN'], ['VERDUN', '55100', 'MEUSE', '55545']],
            [[u'GEANT CASINO', u'MONTPELLIER AVENUE DE LOD\xc8VE'], ['MONTPELLIER', '34000', 'HERAULT', '34172']], # 34070?
            [[u'CARREFOUR', u'EVRY'], ['EVRY', '91000', 'ESSONNE', '91228']],
            [[u'CARREFOUR MARKET', u'ANNEMASSE VERDUN'],  ['ANNEMASSE', '74100', 'HAUTE SAVOIE', '74012']],
            [[u'CARREFOUR MARKET', u'BARBEZIEUX SAINT HILAIRE'], ['BARBEZIEUX ST HILAIRE', '16300', 'CHARENTE', '16028']],
            [[u'INTERMARCHE SUPER', u'PLOUHINEC'], ['PLOUHINEC', '56680', 'MORBIHAN', '56169']],
            [[u'CARREFOUR MARKET', u'AUBUSSON'], ['AUBUSSON', '23200', 'CREUSE', '23008']],
            [[u'LECLERC', u'ST PRIEST'], ['ST PRIEST', '69800', 'RHONE', '69290']],
            [[u'SUPER U', u'CHANGE'], ['CHANGE', '72560', 'SARTHE', '72058']]]

path_source = os.path.join(path_data,
                           'data_supermarkets',
                           'data_source',
                           'data_qlmc_2007-12')

path_source_json = os.path.join(path_source, 'data_json')                          

enc_json(ls_match,
         os.path.join(path_source_json,
                      'ls_city_match.json'))
