==> code_supermarkets_france/data_acquisition/qlmc_2007_2012/build_qlmc_data/1_update_df_stores_w_insee_code.py <==
#!/usr/bin/python
# -*- coding: utf-8 -*-

import add_to_path
from add_to_path import *
from functions_generic_qlmc import *
from functions_geocoding import *
from functions_string import *
from matching_insee import *
import os, sys
import re
import numpy as np
import pandas as pd
from mpl_toolkits.basemap import Basemap
import pprint

path_source = os.path.join(path_data,
                           'data_supermarkets',
                           'data_source',
                           'data_qlmc_2007-12')

path_source_csv = os.path.join(path_source,
                               'data_csv')

path_source_json = os.path.join(path_source,
                                'data_json')

path_match_insee = os.path.join(path_data,
                                u'data_insee',
                                u'match_insee_codes')

# ###########
# LOAD DATA
# ###########

df_raw_stores = pd.read_csv(os.path.join(path_source_csv,
                                         'df_stores_raw.csv'),
                      encoding = 'UTF-8')

#ls_ls_stores = dec_json(os.path.join(path_dir_built_json, 'ls_ls_stores.json'))

df_corr = pd.read_csv(os.path.join(path_match_insee,
                                   'df_corr_gas.csv'),
                      dtype = str)
ls_corr = [list(x) for x in df_corr.to_records(index = False)]
ls_corr = format_correspondence(ls_corr)

# ###############################
# MATCH STORES WITH INSEE CODE
# ###############################

# More complicated than usual: no zip nor dpt => only city name
# of course... city name can be ambiguous hence need to check results
# might need to load recently built class to use format_str_city_insee function
# large part done manually => ls_city_match

ls_rows = []
for i, row_store in df_raw_stores.iterrows():
    chain, city = get_split_chain_city(row_store['store'], ls_chain_brands)
    row = [row_store['period'], chain, city, []]
    city_standardized = format_str_city_insee(city.replace(u"''", u" "))
    for corr_row in ls_corr:
      if city_standardized == corr_row[0]:
        row[3].append(corr_row)
    ls_rows.append(row)

# Check problems and, aside, manually establish a list of corrections
ls_city_match = dec_json(os.path.join(path_source_json,
                                      'ls_city_match.json'))
c = 0
ls_pbms = []
for row in ls_rows:
  if len(row[3]) == 1:
    c+=1
  elif (len(row[3]) > 1) and\
       (all(res[3] == row[3][0][3] for res in row[3])):
    row[3] = [row[3][0]] # what about city name though?
    c+=1
  else:
    pb_bool = True
    for store, row_insee in ls_city_match:
      if row[1:3] == store:
        row[3] = [row_insee]
        pb_bool = False
        break
    if pb_bool == False:
      c+=1
    else:
      ls_pbms.append(row)
print u'Nb no insee match', len(ls_pbms)

# Normally no pbm left in insee matching hence can attribute the only result
for row in ls_rows:
  if len(row[3]) != 1:
    print u'Need to fix', row
    break
  else:
    row[3] = row[3][0]

ls_rows_stores = [row[:3] + list(row[3]) for row in ls_rows]

# BUILD STORE DATAFRAME

ls_columns = ['period', 'store_chain', 'store_municipality',
              'insee_municipality', 'c_insee_zip', 'insee_departement', 'c_insee']
df_stores = pd.DataFrame(ls_rows_stores, columns = ls_columns)

df_stores['store'] = df_stores['store_chain'] + ' ' + df_stores['store_municipality']

df_stores.drop(['insee_departement', 'c_insee_zip'],
               axis = 1,
               inplace = True)

# Precise c_insee_ardt (could have two fields?)

ls_insee_fix = [[[0, u'AUCHAN MARSEILLE'],             [u'marseille', u'13210']],
                [[0, u'CHAMPION PARIS'],               [u'paris', u'75116']],
                [[0, u'E.LECLERC MARSEILLE'],          [u'marseille', u'13209']],
                [[0, u'SYSTEME U LYON'],               [u'lyon', u'69387']],
                [[0, u'CHAMPION MARSEILLE'],           [u'marseille', u'13208']],
                [[0, u'GEANT MARSEILLE'],              [u'marseille', u'13212']],
                [[0, u'GEANT PARIS'],                  [u'paris', u'75113']],
                [[1, u'CARREFOUR LYON'],               [u'lyon', u'69383']],
                [[1, u'SUPER U MARSEILLE RUE TADDEI'], [u'marseille', u'13207']],
                [[1, u'GEANT CASINO PLAN DE CAMPAGNE'],[u'plan de campagne', u'13071']],
                [[1, u'GEANT CASINO MARSEILLE'],       [u'marseille', u'13211']],
                [[2, u'CHAMPION PARIS 13 NATIONALE'],  [u'paris', u'75113']],
                [[2, u'CHAMPION BIHOREL'],             [u'bihorel', u'76108']],
                [[7, u'AUCHAN MARSEILLE'],             [u'marseille', u'13210']],
                [[9, u'CARREFOUR LYON'],               [u'lyon', u'69383']],
                [[9, u'AUCHAN MARSEILLE'],             [u'marseille', u'13210']],
                [[9, u'LECLERC MARSEILLE'],            [u'marseille', u'13209']]]

for [period, store], [insee_municipality, c_insee] in ls_insee_fix:
	df_stores.loc[(df_stores['store'] == store) &\
                 (df_stores['period'] == period),
                ['insee_municipality', 'c_insee']] = insee_municipality, c_insee

# CHECK SUPERMARKET NAMES

## Stores likely to generate multi matching but with disambig. possibilities
## C.C. CITE2 is SOUVENIR FRANCAIS => for matching
#print df_stores[df_stores['INSEE_ZIP'].str.slice(stop=2) == '75'].to_string()
#print df_stores[(df_stores['Commune'].str.contains('carcassonne', case=False)) |\
#                (df_stores['INSEE_Commune'].str.contains('carcassonne', case=False))]
df_stores.sort(['c_insee', 'period', 'store_municipality'], inplace = True)
se_insee_vc = df_stores['c_insee'].value_counts()
# default: sorted in decreasing order
for insee_code in se_insee_vc[0:20].index:
	print '\n\n', df_stores[df_stores['c_insee'] == insee_code].to_string()


# ####################
# LOAD QLMC STORE INFO
# ####################

ls_ls_qlmc_store_info = dec_json(os.path.join(path_source_json,
                                              'ls_ls_qlmc_store_info.json'))

# Store names in qlmc price files
ls_ls_store_names = [list(df_stores['store'][df_stores['period'] == i].values)\
                       for i in range(13)]

# Store names in info files
ls_ls_store_info_names = [[x[0] for x in row] for row in ls_ls_qlmc_store_info]

for i, ls_store_info_names in enumerate(ls_ls_store_info_names):
  print u'\nSize of info file {:d}: {:d}'.format(i, len(ls_store_info_names))
  print u'Matching vs each period price records:'
  for j in range(13):
    set_isct = set(ls_store_info_names).intersection(set(ls_ls_store_names[j]))
    print j, len(set_isct)

# 0th file => Little wit 0-3... check 4-5?
# 1st file => period 6 (100%)
# 2d  file => period 7 (100%)
# 3d  file => period 9 (100% match in price file but got more info: why?)
# 4th file => period 11 (100% match in price file but got more info: why?)
# 5th file => Check that really not period 12?

# Global matching
set_store_info_names = set([store for ls_stores in ls_ls_store_info_names\
                                  for store in ls_stores])
set_store_names = set([store for ls_stores in ls_ls_store_names for store in ls_stores])
# elements in info not in price files : 878
set_uselessinfo = set_store_info_names.difference(set_store_names)
# elements in price file not in info at all (any period)
print u'\nMatching of each price file name vs. all info'
for i, ls_store_names in enumerate(ls_ls_store_names):
  set_isct = set(ls_store_names).intersection(set_store_info_names)
  print i, len(set_isct)

# Check those in info not in price files of periods 9 and 11
set_nomatch_9 = set(ls_ls_store_info_names[3]).difference(set(ls_ls_store_names[9]))
# pprint.pprint(list(set_nomatch_9))
set_nomatch_11 = set(ls_ls_store_info_names[4]).difference(set(ls_ls_store_names[11]))
# pprint.pprint(list(set_nomatch_11))
# For both: essentially hard discount: probably not enough products

# todo: check those missing in price files of period 8, 10 (and 12)
set_nomatch_8 = set(ls_ls_store_names[8]).difference(set_store_info_names)
# pprint.pprint(list(set_nomatch_8))
set_nomatch_10 = set(ls_ls_store_names[10]).difference(set_store_info_names)

# todo: make sure never two stores with same name within info files
ls_rows_qlmc_store_info  = [[i] + store for i, ls_store in enumerate(ls_ls_qlmc_store_info)\
                              for store in ls_store]
df_qlmc_store_info = pd.DataFrame(ls_rows_qlmc_store_info,
                                  columns = ['period',
                                             'store',
                                             'c_qlmc_departement',
                                             'qlmc_surface'])
# print df_qlmc_stores.to_string()

# todo: make sure never two stores with same name / diff dpt or surface across info files
se_qsi_vc = df_qlmc_store_info['store'].value_counts()
#for x in se_qsi_vc[se_qsi_vc > 1].index:
#  print df_qlmc_store_info[df_qlmc_store_info['Magasin'] == x]
## presen of variations in surfaces (maybe precision issue sometimes but also chges)

# Merge store_info with df_stores conservatively i.e. by period
df_qlmc_store_info = df_qlmc_store_info[(df_qlmc_store_info['period'] != 0) &\
                                        (df_qlmc_store_info['period'] != 5)]

ls_replace_periods = [(1,6),
                      (2,7),
                      (3,9),
                      (4,11)]
for store_info_per, qlmc_per in ls_replace_periods:
  df_qlmc_store_info.loc[df_qlmc_store_info['period'] == store_info_per,
                         'period'] = qlmc_per

# Delete duplicate in period 9
df_qlmc_store_info.drop_duplicates(subset = ['period', 'store'], inplace = True)

df_stores_all = pd.merge(df_qlmc_store_info,
                         df_stores,
                         how = 'right',
                         on = ['period', 'store'])
df_stores_all['c_qlmc_departement'] =\
  df_stores_all['c_qlmc_departement'].apply(lambda x: x.rjust(2,'0')\
                                                    if not pd.isnull(x) else x)
# any use?
df_stores_all['period'] = df_stores_all['period'].apply(lambda x: int(x))

# make sure dpt is consistent
print df_stores_all[(~df_stores_all['c_qlmc_departement'].isnull()) &\
                     (df_stores_all['c_qlmc_departement'] !=\
                        df_stores_all['c_insee'].str.slice(stop = 2))].to_string()

ls_final_insee_fix = [[[9, u'CARREFOUR MARKET ROMILLY'], [u'romilly sur seine', '10323']],
                      [[9, u'CARREFOUR MARKET ST OUEN'], [u'saint ouen', '41126']],
                      [[9, u'LECLERC ROMILLY'],          [u'romilly sur seine', '10323']]]

for [period, store], [insee_municipality, c_insee] in ls_final_insee_fix:
	df_stores_all.loc[(df_stores_all['store'] == store) &\
                    (df_stores_all['period'] == period),
                    ['insee_municipality', 'c_insee']] = insee_municipality, c_insee

df_stores_all.drop(['c_qlmc_departement'],
                   axis = 1,
                   inplace = True)

# #######
# OUTPUT
# #######

# CSV
df_stores_all.to_csv(os.path.join(path_source_csv,
                                  'df_stores_w_municipality.csv'),
                     index = False,
                     encoding = 'UTF-8')
